my initial commit with a map, layers, and buttons that move a marker
