+++ src/border.js
@@ -1,98 +1,18 @@
-const borderData = {
-  "type": "Feature",
-  "properties": {
-    "name": "Vermont",
-    "name_alt": "VT",
-    "type": "State",
-    "latitude": 44.0886,
-    "longitude": -72.7317
-  },
-  "geometry": {
-    "type": "Polygon",
-    "coordinates": [
-      [
-        [-73.35218221090553, 45.00541896831666],
-        [-73.18201496808604, 45.00562770855899],
-        [-72.76588580182134, 45.00611110701493],
-        [-72.34976762188519, 45.006605491799434],
-        [-71.9336384556205, 45.00706691759828],
-        [-71.51752027568435, 45.007561302382754],
-        [-71.53348341105892, 44.98797267858865],
-        [-71.51022535353107, 44.90834376930104],
-        [-71.62065992805631, 44.77189356878338],
-        [-71.6182868810908, 44.72776148702175],
-        [-71.56844190848624, 44.60763697072062],
-        [-71.57149610782149, 44.57919336611994],
-        [-71.58779981938093, 44.565262701526024],
-        [-71.60928907801318, 44.514055424182004],
-        [-71.63820509474124, 44.481678713962566],
-        [-71.68300734254406, 44.45027978698354],
-        [-71.82551101008951, 44.37408959853025],
-        [-71.86767653904175, 44.3549184552208],
-        [-72.00177566525097, 44.32949609096974],
-        [-72.0312080394206, 44.300733882841314],
-        [-72.06223343122913, 44.11635033088575],
-        [-72.10621170439114, 44.013078842571275],
-        [-72.11492386292662, 43.965431135675516],
-        [-72.17385452923715, 43.884088359135035],
-        [-72.22238114241641, 43.79086936143841],
-        [-72.29669266868866, 43.7149538311987],
-        [-72.33767167415814, 43.621987519058564],
-        [-72.36236894072528, 43.58664450013222],
-        [-72.38495683221193, 43.52922994716076],
-        [-72.4070723115712, 43.3320363361228],
-        [-72.43818559400808, 43.224644974604274],
-        [-72.47371538051968, 43.038536569067304],
-        [-72.49647905326304, 42.99166889149819],
-        [-72.51941850726311, 42.966674994060384],
-        [-72.54980669201608, 42.886683535930814],
-        [-72.55280595970862, 42.85644915977747],
-        [-72.5396223654557, 42.82963153180134],
-        [-72.48365800785209, 42.76617449813065],
-        [-72.46685991150817, 42.730315121762715],
-        [-72.56517656564927, 42.733050717570194],
-        [-72.66349321979038, 42.73578631337767],
-        [-72.76177691494586, 42.738532895513686],
-        [-72.86009356908696, 42.741268491321165],
-        [-72.95841022322807, 42.744004087128644],
-        [-73.05669391838354, 42.74673968293612],
-        [-73.15501057252465, 42.749486265072164],
-        [-73.25332722666576, 42.752221860879644],
-        [-73.28019978628461, 42.81315203898518],
-        [-73.26655476623284, 42.864776796813885],
-        [-73.26454426810928, 42.93521014911008],
-        [-73.26192952191579, 43.027066842067256],
-        [-73.25908406282286, 43.1278554201308],
-        [-73.25523884783243, 43.26537129451725],
-        [-73.25265706062457, 43.357195028488775],
-        [-73.25032795897322, 43.439669396869306],
-        [-73.24709797838126, 43.55304830744436],
-        [-73.3061604806343, 43.616252655558554],
-        [-73.33048421203094, 43.626689667675436],
-        [-73.3492378748557, 43.620383515091135],
-        [-73.39143636279358, 43.578525603338136],
-        [-73.40494954690281, 43.57881124788028],
-        [-73.41592488911837, 43.589643767824754],
-        [-73.4133650745676, 43.62785421850111],
-        [-73.3758467625895, 43.76218405760977],
-        [-73.37496785630599, 43.804151832648216],
-        [-73.38300984880026, 43.875826640069874],
-        [-73.42494466485307, 44.038292466579975],
-        [-73.42613118833583, 44.074854967974716],
-        [-73.40975057247658, 44.131412587319716],
-        [-73.38112020029067, 44.18743187656605],
-        [-73.33182454411332, 44.24458275765244],
-        [-73.32091511986903, 44.268873530063445],
-        [-73.32463948524548, 44.367860350245735],
-        [-73.30975301006823, 44.4596401389031],
-        [-73.37207845189889, 44.59733179454625],
-        [-73.34856770881451, 44.77529933063204],
-        [-73.3658052583002, 44.860366472548975],
-        [-73.34505208368041, 44.93886378999653],
-        [-73.35218221090553, 45.00541896831666]
-      ]
-    ]
-  }
-}
-
-export default borderData;
+let countyData = {"type":"FeatureCollection", "features": [
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-71.89930556685702,45.007967038225665],[-71.90639495793083,44.972034737155035],[-71.91256962807861,44.941586303684424],[-71.92014649911405,44.901534959428105],[-71.88768224564625,44.88414402607262],[-71.87785847130161,44.87877145078154],[-71.89809815971485,44.85887096608518],[-71.91995085333807,44.83690375089462],[-71.92931744591431,44.82964832537888],[-71.93683910186412,44.822541420493785],[-71.97482609720078,44.78724062557453],[-71.93929273719223,44.76945013143421],[-71.95140936345551,44.757495286499],[-71.9750836685309,44.734647555007385],[-71.99123898495328,44.71882667502445],[-71.9995389566063,44.711086418730154],[-72.01145503900082,44.698841717258844],[-72.05216258622916,44.718994083231216],[-72.08972446262909,44.73767356435365],[-72.11436109432913,44.74982120146015],[-72.11197436534691,44.74320799587801],[-72.10194093000278,44.71773911466352],[-72.08929717635908,44.68509008796579],[-72.12452275235108,44.69729842843127],[-72.1459577724481,44.704507419885516],[-72.14996003178632,44.6978650872916],[-72.18495758607976,44.64925658454629],[-72.19926218787573,44.628291149617525],[-72.2024891506249,44.62410595577774],[-72.20712445792117,44.6167421524504],[-72.22889633382701,44.58798400220501],[-72.2616720312797,44.54376984071134],[-72.26289163479404,44.54250346778505],[-72.28771378786477,44.55181867632704],[-72.30319817668146,44.55778858931625],[-72.35419116580978,44.57644905463246],[-72.36653172204848,44.58079434275957],[-72.37386153254992,44.584171294492144],[-72.37175984409885,44.58823645937912],[-72.36838584940472,44.59285547005072],[-72.37112518348113,44.593344555644464],[-72.45815134995883,44.62505021122964],[-72.48279415530956,44.634112332199315],[-72.45420358363344,44.67313270642213],[-72.41767326776278,44.722688949026555],[-72.4224471271231,44.72447131979226],[-72.50087300480874,44.752627096688336],[-72.52019325610684,44.75974616084036],[-72.57907591571808,44.780904784770726],[-72.53328173402815,44.834089982270065],[-72.53621528661253,44.86799211898105],[-72.53698928285543,44.8750039568626],[-72.54138462174261,44.92598662813111],[-72.53650006984168,44.92612408518761],[-72.53850917778888,44.935403935366374],[-72.55123785454646,44.99658426579739],[-72.55248533250622,45.00300225609495],[-72.55392229303604,45.00813729074689],[-72.53091991438157,45.00778039262025],[-72.47902537887512,45.008838841381674],[-72.45043879964273,45.00857176071912],[-72.42484626196176,45.00775581214537],[-72.420470493147,45.00747246941556],[-72.38800538212027,45.00626980117705],[-72.36863243694785,45.006225410720376],[-72.32666231926974,45.00481523852764],[-72.314255492748,45.00385615373037],[-72.2933492976557,45.00436457144448],[-72.27021386211268,45.004208558100245],[-72.25564535743473,45.004431393922964],[-72.24283281090447,45.004873748680374],[-72.2314723708005,45.00489811566134],[-72.21063837212338,45.005300925934165],[-72.17302884110566,45.005850986351746],[-72.1510671874543,45.00603159699963],[-72.11041075734528,45.00562616785874],[-72.05428436960499,45.00622906217937],[-71.94575433608884,45.00837926498306],[-71.92932910820353,45.00811542418737],[-71.91398167725643,45.007637693207535],[-71.89930556685702,45.007967038225665]]]},"properties":{"OBJECTID":1,"CNTY":19,"CNTYNAME":"ORLEANS","ShapeSTArea":1870297146.3263016,"ShapeSTLength":209703.88946562487}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-73.36204303000723,44.56446315634094],[-73.37074740354149,44.57250271504769],[-73.37457216775015,44.57568984118599],[-73.3755084407029,44.5817076079403],[-73.37785429508112,44.58538713887643],[-73.38171484654292,44.58960956071945],[-73.38082306910394,44.59155929685536],[-73.37725574778058,44.59457001515015],[-73.37630988815145,44.59754523674588],[-73.37723427758299,44.60049259939323],[-73.37959508126276,44.60332621720652],[-73.38182558039327,44.60838090706059],[-73.38226463623955,44.611294982997386],[-73.38364157068257,44.612943888387136],[-73.38929668282083,44.6167774955149],[-73.39019005349404,44.61894205137402],[-73.38732210212889,44.62373576367784],[-73.38648038124136,44.62688384046522],[-73.38606048650858,44.631919372931954],[-73.3870527812515,44.63591172987294],[-73.38554562661518,44.63740289658006],[-73.37884875253079,44.64126948228766],[-73.3800298500661,44.64300524192905],[-73.38438148451885,44.64676730652772],[-73.3798066276536,44.650411318287674],[-73.37802789411523,44.65302406266602],[-73.37891400195753,44.65704257996462],[-73.37691262984491,44.65841864089187],[-73.37440213057718,44.66222645230369],[-73.37033128337248,44.66278049073751],[-73.36996075034467,44.66570156377635],[-73.37276699497235,44.669044938448806],[-73.37112372635217,44.67293783132166],[-73.37193312131588,44.67709916074956],[-73.36731471418831,44.67862629777496],[-73.36736900365493,44.68102226716738],[-73.37007500264356,44.68497660178258],[-73.3661645074577,44.68706362014728],[-73.36218293678708,44.692830692986476],[-73.36127980385515,44.69498730788902],[-73.36461975183914,44.69637655107657],[-73.3659852153364,44.69798973535462],[-73.36538203882185,44.70348997443884],[-73.36580444621825,44.707160116734684],[-73.36521803412111,44.71073450318694],[-73.36557132761638,44.718535745903914],[-73.36504517895557,44.72572981545355],[-73.36511969217595,44.73530947497335],[-73.36562002703869,44.74181631766825],[-73.36159708842175,44.74773580377236],[-73.35774524171642,44.75081367313177],[-73.35423296687972,44.7556057958136],[-73.3514018156148,44.763090815656646],[-73.34906655558251,44.76694637053321],[-73.34667591449596,44.773672786862804],[-73.34524077364955,44.775424813937036],[-73.33596206870024,44.78203267963672],[-73.33347347629,44.78621838825722],[-73.3331990434045,44.78970740416255],[-73.33352089580409,44.79834496350986],[-73.3351746910078,44.80426580206339],[-73.34142351927822,44.80875891011707],[-73.3436332757872,44.81157259418567],[-73.35024934372825,44.81638583147413],[-73.35504981556959,44.82152320274236],[-73.3610699314474,44.824624891498395],[-73.36510260125577,44.82609633259585],[-73.36964430653707,44.82912209732695],[-73.37455010397201,44.835502743262786],[-73.37973749088482,44.8386555286464],[-73.3795093219521,44.84358499003538],[-73.3813563661971,44.844899075634586],[-73.3804274648697,44.84756811122877],[-73.38158930275142,44.849232039992266],[-73.37947884981646,44.851839900021375],[-73.37985085704307,44.85679640316842],[-73.37785403402528,44.85792934002066],[-73.37604604347989,44.860406303497385],[-73.371707044177,44.86266672535446],[-73.36979327474779,44.86544846658627],[-73.3680802299451,44.86944799301252],[-73.36507868489988,44.881188070269836],[-73.36075479434646,44.89662852195265],[-73.35735093273448,44.90279941009156],[-73.35483014052008,44.90640882677928],[-73.34824518062715,44.911256263176],[-73.34341909065164,44.913338152237614],[-73.33970150351696,44.916389408308596],[-73.33850165967546,44.92175501416426],[-73.33925516584506,44.93227181997692],[-73.33961641075724,44.945950671973215],[-73.338020575839,44.96099664503955],[-73.33818823707988,44.96480735607246],[-73.34009495498009,44.967113210213085],[-73.34470756148079,44.97053693338275],[-73.34845113599371,44.97397565683505],[-73.35215856205063,44.97903482782102],[-73.35424903705098,44.983979459628976],[-73.35476296159045,44.986804526068354],[-73.35413249086253,44.989280277098835],[-73.35076718841873,44.99393997821094],[-73.34952544700772,44.997927437301264],[-73.34293440428617,45.01072501288562],[-73.27237164183947,45.01207160423754],[-73.2263408268779,45.01276625006285],[-73.19231232645926,45.013479954714725],[-73.19635850423604,45.0057668319345],[-73.20106588489008,44.99827032524488],[-73.20699129309702,44.98961389184801],[-73.21629732083638,44.97775944966089],[-73.22010863695098,44.97258913132462],[-73.22326801487539,44.96703856753372],[-73.22896608209605,44.959988746843905],[-73.23913032448247,44.9532238696633],[-73.24094565413421,44.94889469800273],[-73.24041127052149,44.945276984819074],[-73.23795114352099,44.94064758060852],[-73.23425103956849,44.93701713089356],[-73.23244529070674,44.93066040061028],[-73.22612632272815,44.91960725686122],[-73.22377659965154,44.91785139428439],[-73.21594625228322,44.91526321129958],[-73.21164889382469,44.913098058373386],[-73.20569845372954,44.91110575813656],[-73.19534916916606,44.910591070524134],[-73.19231092187229,44.908302150006776],[-73.19131868889878,44.90561221238166],[-73.19144757967253,44.89807970533099],[-73.19199539609963,44.89252583022596],[-73.19304145672403,44.88894164653275],[-73.19567185557246,44.882978239113385],[-73.19643351387317,44.87542888221953],[-73.19830643782595,44.871848503069764],[-73.20333236769989,44.865995492476664],[-73.20351034473015,44.861803716947705],[-73.20705637918306,44.85195675950775],[-73.20755552328397,44.846104393775],[-73.20676654949014,44.843867745886854],[-73.20311722066053,44.83853502987461],[-73.20322210104487,44.8336399703634],[-73.20704826571769,44.8292256868831],[-73.20757657601354,44.82554374737249],[-73.210821451737,44.82210130594112],[-73.21687522591805,44.82062066681408],[-73.22412968111153,44.81815859501493],[-73.23006310953835,44.81469483351476],[-73.23703144649159,44.8096626806059],[-73.24291410052965,44.803326018684984],[-73.24574750976059,44.799588372202614],[-73.2526901120541,44.78901853156588],[-73.25591393442753,44.7831710278447],[-73.2588871386311,44.774096792880485],[-73.2588370531649,44.76809200039521],[-73.25797255202558,44.76359553580399],[-73.25529863079942,44.758081554348706],[-73.25013579223629,44.750224603789306],[-73.22885687298746,44.722817895851385],[-73.2296384288906,44.71316426386114],[-73.22940104582004,44.70811107655741],[-73.22854139129336,44.70306905503584],[-73.22840732921972,44.68827094414793],[-73.22891141469866,44.68562468177842],[-73.23273099314,44.67512073389787],[-73.2335626675656,44.67160392393109],[-73.23287228541838,44.65143427587074],[-73.23384006515026,44.648390023726336],[-73.23770869231676,44.64590183928723],[-73.24275462298844,44.64401970754641],[-73.24951566840585,44.64241327354453],[-73.25111470657615,44.641520845840354],[-73.25485872804511,44.635006739435],[-73.25628351220969,44.629685973620674],[-73.25586157208194,44.62263535104321],[-73.25479833752622,44.61534327910245],[-73.2545038030464,44.61068541379099],[-73.25513211622983,44.60682858939562],[-73.25734054352043,44.60269680314303],[-73.26194646009395,44.597541262509644],[-73.26985345777844,44.592894914505344],[-73.27932110503008,44.58860274763236],[-73.2892685564037,44.58463873192691],[-73.3031319007977,44.57974725652776],[-73.32766297480886,44.572804453735465],[-73.33877091044171,44.5703511941668],[-73.3508544865256,44.56710664808632],[-73.36204303000723,44.56446315634094]]]},"properties":{"OBJECTID":2,"CNTY":13,"CNTYNAME":"GRAND ISLE","ShapeSTArea":500854749.3310623,"ShapeSTLength":130415.47682757318}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-72.92377886089447,44.6332237262317],[-72.90996235025344,44.63120631512025],[-72.87919405921666,44.60071893243115],[-72.85846799278407,44.579787270846],[-72.82704363164359,44.55070069488409],[-72.80995811259517,44.54672310649224],[-72.83079748289437,44.51643552620223],[-72.80362994552533,44.5062898887272],[-72.8382972186703,44.45930183682008],[-72.80320710340565,44.4515249990348],[-72.81435655050836,44.43675183361455],[-72.83331806207146,44.4118588461083],[-72.83983118375824,44.40346104201913],[-72.80237685381812,44.38858937755037],[-72.82595844086234,44.35903495682558],[-72.82875852758922,44.36037562087989],[-72.83283347714442,44.358779969756306],[-72.83588657792055,44.360181909221275],[-72.84382894658889,44.36221595927199],[-72.84935519174762,44.36297172119535],[-72.85555099714806,44.36559517079108],[-72.86125327167196,44.36690949653793],[-72.88039671983401,44.32986246096231],[-72.90595977390268,44.28056333642762],[-72.89498176046104,44.27810636324729],[-72.90208280627773,44.26361270517944],[-72.91757928190204,44.23238366831349],[-72.94199936815059,44.18246161980573],[-72.95220573089938,44.16142046923946],[-72.95703444625443,44.185816672730425],[-72.96545080319989,44.23139972811686],[-72.97080187807347,44.26334646181387],[-72.97588174197186,44.29504739217094],[-73.0058479191561,44.3009831930459],[-73.02037284201288,44.30375376633077],[-73.01651933434506,44.28536585557466],[-73.07001238401223,44.279934866078236],[-73.07058802229065,44.2838702930658],[-73.08439116483443,44.28171859313689],[-73.10473136432508,44.278842321269124],[-73.15219676486767,44.27245979434792],[-73.16183248633514,44.271536612655844],[-73.19679836287507,44.266642030713605],[-73.24914345799375,44.26177231255189],[-73.27038317174949,44.26011511493143],[-73.31328149172086,44.26414772030446],[-73.31121275305527,44.27098426677069],[-73.31098480822975,44.27649563704971],[-73.31176585881111,44.27919901909162],[-73.31663655660742,44.28759140470035],[-73.31939319563682,44.29290941362478],[-73.32185660363892,44.30049082426265],[-73.32372541109301,44.30883090767313],[-73.32464064555417,44.31806151782783],[-73.32405559631289,44.33322032298268],[-73.32444048186915,44.336268256633645],[-73.32718208788278,44.34335916468316],[-73.33398132856198,44.354510424138404],[-73.33542815979459,44.359946344883745],[-73.33493250742264,44.36946527980421],[-73.33408325311986,44.37274563715015],[-73.33108141597766,44.37568600479158],[-73.32523224640461,44.38007801789645],[-73.32018049619724,44.384812199563704],[-73.31696147836303,44.389628365500755],[-73.31633372363291,44.394524131286836],[-73.31467149778251,44.40063951600069],[-73.31216994986407,44.4059400809525],[-73.30648504894484,44.41526890204086],[-73.29816725957397,44.42520424550898],[-73.29599950255161,44.42874350271463],[-73.29417257188838,44.433698672907305],[-73.29450274308259,44.43602275383971],[-73.2934009162084,44.43750732472701],[-73.29401005360904,44.44381805204804],[-73.29635248830495,44.448723393922414],[-73.29937366340641,44.45365561321203],[-73.2997002595546,44.45782009187172],[-73.29821706482518,44.46562293988749],[-73.29838330960882,44.46940155704949],[-73.29963496869655,44.477457131140426],[-73.30346037654132,44.484463356503795],[-73.3051330777482,44.49286146464182],[-73.30790156365107,44.49966879136842],[-73.31129178140876,44.505040721115876],[-73.31722828105664,44.51087312178481],[-73.31941720444249,44.513633753909474],[-73.3203714220553,44.51727166020202],[-73.31941227391633,44.52203193492885],[-73.3214609058837,44.52459185269402],[-73.3276059523952,44.52702681787352],[-73.3321585357475,44.53081298264569],[-73.33523629428306,44.53533321792193],[-73.3380368555229,44.54039914230917],[-73.33937565262383,44.54487989915775],[-73.3393922908932,44.54839303304007],[-73.34919521241812,44.55435624161939],[-73.35121854208242,44.55516731802226],[-73.35954595558731,44.561534878951186],[-73.36204303000723,44.56446315634094],[-73.3508544865256,44.56710664808632],[-73.33877091044171,44.5703511941668],[-73.32766297480886,44.572804453735465],[-73.3031319007977,44.57974725652776],[-73.2892685564037,44.58463873192691],[-73.27932110503008,44.58860274763236],[-73.26985345777844,44.592894914505344],[-73.26194646009395,44.597541262509644],[-73.25734054352043,44.60269680314303],[-73.25513211622983,44.60682858939562],[-73.2545038030464,44.61068541379099],[-73.25479833752622,44.61534327910245],[-73.25586157208194,44.62263535104321],[-73.25628351220969,44.629685973620674],[-73.25485872804511,44.635006739435],[-73.25111470657615,44.641520845840354],[-73.24951566840585,44.64241327354453],[-73.24275462298844,44.64401970754641],[-73.23770869231676,44.64590183928723],[-73.23384006515026,44.648390023726336],[-73.23287228541838,44.65143427587074],[-73.2335626675656,44.67160392393109],[-73.23273099314,44.67512073389787],[-73.22891141469866,44.68562468177842],[-73.22840732921972,44.68827094414793],[-73.22854139129336,44.70306905503584],[-73.22940104582004,44.70811107655741],[-73.2296384288906,44.71316426386114],[-73.22885687298746,44.722817895851385],[-73.16068376809086,44.69751869943463],[-73.10282514170754,44.67616487782963],[-73.03807607862669,44.65115984590366],[-73.03783661824305,44.65129047227957],[-73.00929421259129,44.646511987647436],[-72.99432691552323,44.64436610715115],[-72.98352770589622,44.642584509790964],[-72.94253298456528,44.636115862088666],[-72.92377886089447,44.6332237262317]]]},"properties":{"OBJECTID":3,"CNTY":7,"CNTYNAME":"CHITTENDEN","ShapeSTArea":1609236731.7965088,"ShapeSTLength":201814.7092398412}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-72.20540077140059,43.771330595300675],[-72.21060785066145,43.768069393217274],[-72.21783859555389,43.76612939387684],[-72.22075101226628,43.76284232642897],[-72.22443965835646,43.757409692529336],[-72.22614388950973,43.75593732732025],[-72.23226634201869,43.747721847111485],[-72.23876807420073,43.74456660123276],[-72.24685960964226,43.742721945172654],[-72.25203407362463,43.73944814627557],[-72.26446365495747,43.73388584429536],[-72.2690176546504,43.73432329950898],[-72.27194190723041,43.733247293418245],[-72.27553098027859,43.72719146691718],[-72.28147814247615,43.72394007020837],[-72.28436140702999,43.72111258683026],[-72.28646126431201,43.7174702665691],[-72.29151972465066,43.71265326035753],[-72.29227681423123,43.71102470565864],[-72.29508128266302,43.708828753069085],[-72.29933154673783,43.70670311537835],[-72.29994189163865,43.7037102840319],[-72.30251220096399,43.69826536964629],[-72.3051255209525,43.695368385471795],[-72.30545005800568,43.68668864626189],[-72.30597498293935,43.68325738117508],[-72.30280168121622,43.67969630064974],[-72.30251563305873,43.6743837512465],[-72.3034536850835,43.668056491445945],[-72.30534457609168,43.667370720080925],[-72.30618765735518,43.66515652451746],[-72.31076340254147,43.65971855018158],[-72.3134628818235,43.657352184432284],[-72.31434475475014,43.655039160430036],[-72.31509670847026,43.64906917303818],[-72.31377029245763,43.64436698857205],[-72.31439718668445,43.6421999870409],[-72.31724639766324,43.6401755287337],[-72.32279630594164,43.638731537247615],[-72.32863574513154,43.635962701020865],[-72.32965650903682,43.634306188029846],[-72.32742981281083,43.631108731816006],[-72.32905151817677,43.627220385664856],[-72.33247929484347,43.62505833145189],[-72.33467514716656,43.61931621017326],[-72.33496180560854,43.614979571135756],[-72.33290218394642,43.60998120274072],[-72.32839173734511,43.6070363742808],[-72.32832541951655,43.60130371908561],[-72.33211019016704,43.599849310972886],[-72.34320307506012,43.5918646406897],[-72.34672089275297,43.5904137003205],[-72.34931086531512,43.58813991998653],[-72.35825097494472,43.585195194522036],[-72.36640037566607,43.582869455757674],[-72.37102505338056,43.580877829497794],[-72.37474740379857,43.57861689196262],[-72.38034792178567,43.57363414202411],[-72.38063670301817,43.56995811321934],[-72.3797608424756,43.567234959476906],[-72.38046515061687,43.565125793191925],[-72.38303896136074,43.56403781377407],[-72.38158586587777,43.555335932346495],[-72.38098285954464,43.54956795517172],[-72.38168556500868,43.54599935956534],[-72.38066424033978,43.54302438407106],[-72.38114764090597,43.53926662934105],[-72.38390639025477,43.534649268721616],[-72.38800763347304,43.52954503953608],[-72.39005091130578,43.52793152748107],[-72.39438981121093,43.52749654882533],[-72.39613789770279,43.52374510205545],[-72.39634207085263,43.51922382051262],[-72.39910221728954,43.51252234659248],[-72.3969140457987,43.50844744110632],[-72.38633172438476,43.50187675369875],[-72.38361892998239,43.498382077886895],[-72.38314281148135,43.49656455539623],[-72.38053747659315,43.49270861250499],[-72.38019010912811,43.49055898124068],[-72.38202783366043,43.478635677121666],[-72.3829572183633,43.475758005032155],[-72.39269803346176,43.46722248722059],[-72.3915615244909,43.4567371358098],[-72.39034125869821,43.45255253429318],[-72.39130390296305,43.44859432166979],[-72.3935434990692,43.44508527559357],[-72.39595877868405,43.43758966157071],[-72.39610436068602,43.42837660031873],[-72.39982630712463,43.41606761612066],[-72.40027172154245,43.40837552322654],[-72.4021792966762,43.39912270440912],[-72.40356868095466,43.39630741511698],[-72.40379757931547,43.392066067862004],[-72.40494330219367,43.39037723186834],[-72.41189785951839,43.3852820914073],[-72.41480243623342,43.38141953429139],[-72.41607601661978,43.37568661301206],[-72.41455886427953,43.36917589524882],[-72.41460005484223,43.36426458379361],[-72.41249890620158,43.36235905391011],[-72.40883700106481,43.3611560585049],[-72.40379532100917,43.357898879611746],[-72.40113885645317,43.35737850260901],[-72.39278802056356,43.35794795954307],[-72.39004231008231,43.35573374550091],[-72.39473057087038,43.35067423253292],[-72.3982933718493,43.348485289565026],[-72.40073679728931,43.34604413349993],[-72.40504591819099,43.339561927154364],[-72.40813262670525,43.33576707890912],[-72.41055406250267,43.33097302364248],[-72.40936287697295,43.32851923432535],[-72.40241948567056,43.32008344520219],[-72.39732695222175,43.31695940415182],[-72.39560469315632,43.31419156190947],[-72.39626842277427,43.311490600528565],[-72.39915035727535,43.308159448245185],[-72.39977227584345,43.30601479886764],[-72.40289834805674,43.302412012306604],[-72.40267914336827,43.3009872748162],[-72.40803885772628,43.28412714173503],[-72.4162035371647,43.27024393473632],[-72.4207415027183,43.26577041853194],[-72.42183517050171,43.26326608321626],[-72.4249750000892,43.26176324196357],[-72.43363559420794,43.25900966477465],[-72.4370012166186,43.256058053776876],[-72.43902298830001,43.25138207923467],[-72.43841246627866,43.249271910261776],[-72.43959089037315,43.24544554220797],[-72.43822867351156,43.23946457799815],[-72.43577071689305,43.234036077712275],[-72.43477237650221,43.23326480936767],[-72.4588539788752,43.2337682936334],[-72.4873383441494,43.23465704295029],[-72.5491533490645,43.235905583761415],[-72.60292208462033,43.23076809664846],[-72.67624428970716,43.22409200171442],[-72.67631750251499,43.22267352942886],[-72.68544184708287,43.22258898153981],[-72.70141356956857,43.226337546532754],[-72.75682319827382,43.239694705034054],[-72.81962056340674,43.25601438201351],[-72.82003172808133,43.2918325660342],[-72.81989655729153,43.2993807446879],[-72.84956948656098,43.30018783120005],[-72.86864799426148,43.300537763988615],[-72.86497354118751,43.348700497881545],[-72.85470147202362,43.349286212713345],[-72.85238506315532,43.37628742808133],[-72.82198340435433,43.375719072315235],[-72.82193521746565,43.380605422746356],[-72.80059721691116,43.37143032610788],[-72.78361960585687,43.36392427189571],[-72.7584914379173,43.35331542836739],[-72.762356028391,43.37143029541823],[-72.75522264269321,43.384725863370484],[-72.73840418568807,43.4170448480358],[-72.72256418430345,43.46008526258974],[-72.72736822235086,43.46107089244957],[-72.72422431737083,43.46818253753719],[-72.75402169880877,43.474900982084804],[-72.77474320844217,43.47936656459434],[-72.77254451909191,43.49220956164092],[-72.79172429123342,43.4970631502777],[-72.78585866267535,43.5249825158868],[-72.78222425102138,43.54323446733776],[-72.77307441914257,43.591446795356354],[-72.75794942164497,43.5874325129525],[-72.75153983430204,43.59801341265904],[-72.7305049194344,43.63074079085405],[-72.72756494419221,43.63618145870498],[-72.6997966587968,43.68458556729994],[-72.81998614127639,43.71670554625107],[-72.80392886112465,43.74098192214516],[-72.81926310468323,43.74632159535934],[-72.79782246707592,43.77938827783577],[-72.78285746796037,43.80401599297043],[-72.7875111310777,43.805715292370635],[-72.78869832705348,43.80730809401689],[-72.787848756436,43.8089681345466],[-72.7843516385832,43.81109282292592],[-72.7844032084616,43.81384073436037],[-72.78853304424172,43.81606233136843],[-72.79211903753617,43.81651044799379],[-72.79423967863556,43.8203221256834],[-72.79372599007979,43.82206234543868],[-72.79129210190986,43.82260320501957],[-72.79541088059247,43.82563480331924],[-72.79692532178815,43.8291670848141],[-72.82238830706459,43.83646013277515],[-72.83066668842413,43.839128524018825],[-72.83834275094526,43.82926312363365],[-72.86138297446533,43.83623700187526],[-72.87387230912563,43.839637062195244],[-72.87449918314891,43.8580692509895],[-72.89147479498777,43.851572609658845],[-72.91037961632605,43.8445630599101],[-72.92723289471468,43.838732715703124],[-72.92847207760501,43.837960335777495],[-72.95974533334999,43.826085702498936],[-72.96551127410675,43.844602358821994],[-72.97289506721809,43.86677686576626],[-72.98023112484987,43.86397512998485],[-72.98015118570738,43.88293286757756],[-72.97404303754013,43.881447290944294],[-72.96974979784696,43.87967762868503],[-72.9531899187358,43.87486735658826],[-72.9454187973384,43.8775584334656],[-72.92762592253814,43.88432468484837],[-72.9027925912886,43.89283217705257],[-72.87655389178019,43.902772931878175],[-72.87448302748876,43.872379171063066],[-72.87534563607142,43.86770511840981],[-72.87417710995096,43.86407016432933],[-72.86529107810452,43.8616561942865],[-72.86127932847306,43.86774793317464],[-72.83831069591231,43.90376684828651],[-72.81632926346374,43.937815227371466],[-72.80999702540517,43.93609650849612],[-72.80350464771345,43.945672456924875],[-72.79104452787709,43.96225273388067],[-72.78705052498368,43.960867072488284],[-72.7917776944291,43.95397707450101],[-72.78522103663506,43.95219986860384],[-72.7899604524742,43.9453235997332],[-72.77108017968231,43.939073531628715],[-72.77514531906822,43.9336668795571],[-72.76891266149019,43.931625740039976],[-72.77350320973747,43.92454741962046],[-72.71228590372552,43.90797268103477],[-72.70930742816792,43.911252309986345],[-72.65388603721946,43.89564164550551],[-72.60255996488932,43.88178874793391],[-72.57359657753446,43.87312611162169],[-72.5453764509597,43.86434935310264],[-72.5045221614655,43.85191469303779],[-72.4595166251448,43.83810151237467],[-72.46156601403018,43.83686015250722],[-72.43831424337422,43.83228950849945],[-72.40506710826833,43.82546657927705],[-72.3461038994287,43.81160567984825],[-72.33934470947001,43.81045547150489],[-72.30512109058216,43.80048346869359],[-72.28720048632498,43.79516098497731],[-72.20846523554275,43.77216821707658],[-72.20540077140059,43.771330595300675]]]},"properties":{"OBJECTID":4,"CNTY":27,"CNTYNAME":"WINDSOR","ShapeSTArea":2529817960.736702,"ShapeSTLength":311640.03352650645}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-72.43477237650221,43.23326480936767],[-72.43481310345408,43.23005113557465],[-72.43951867145178,43.22260168618121],[-72.44063111357423,43.22021391188691],[-72.44125494063606,43.21619300858676],[-72.43884727138912,43.20839511837971],[-72.43906381609646,43.20101482372001],[-72.44560033065733,43.19569904295527],[-72.4504278331102,43.19290869956159],[-72.44968935898142,43.18953731187952],[-72.44780425203106,43.18657352890324],[-72.4446860284134,43.183483462606],[-72.44388549445883,43.18052058119689],[-72.44436705619239,43.17861762329927],[-72.45100020879015,43.17484901407071],[-72.45294509271255,43.171939362305096],[-72.45181549665877,43.1651772922372],[-72.451708752569,43.161277954145824],[-72.4524783733152,43.15859815307718],[-72.45224882896089,43.152887553375365],[-72.45662059323836,43.14940856600066],[-72.45739289683974,43.14747684855126],[-72.4530596012163,43.13947688374008],[-72.45108557740245,43.137990074054976],[-72.44720475704135,43.13692220139206],[-72.44398372432768,43.137632875223694],[-72.44104908411227,43.13673772356756],[-72.4403069652453,43.13429448351514],[-72.44221035075323,43.131944694220905],[-72.44218063201737,43.12944097185973],[-72.44052661447843,43.12718171052132],[-72.43707554802297,43.124689298692985],[-72.43284120793005,43.11941815452848],[-72.43248809584539,43.115464072572365],[-72.43411971750929,43.11019627669005],[-72.43684505403978,43.10827128424282],[-72.4418154799131,43.105852386643804],[-72.4435038301412,43.10200295797394],[-72.44211108063881,43.09825053392003],[-72.43897623813675,43.094823522818885],[-72.43906464887236,43.09175409337108],[-72.43820962821607,43.08995654101324],[-72.4336621374078,43.085769477766185],[-72.43683053224935,43.079612230460036],[-72.43960347234766,43.07651156566422],[-72.4461432338094,43.07051889076185],[-72.4503964282393,43.068535037801816],[-72.45548877292556,43.062630570857564],[-72.46188595980766,43.05898583492817],[-72.46572852080372,43.055786667249464],[-72.46789854120493,43.052746874378],[-72.46724244111537,43.0488563192674],[-72.46536449761979,43.04607810138391],[-72.46159932051717,43.0432337010295],[-72.4606672116865,43.04091120339112],[-72.4615690634312,43.03331385014325],[-72.46322764796105,43.028488460821755],[-72.46285603874352,43.025767195232575],[-72.45805930095399,43.01825301610651],[-72.45235271205858,43.01533852445242],[-72.44377790830039,43.009942560658075],[-72.44320854694786,43.0089722459838],[-72.44407910616948,43.0052756352504],[-72.44641905535023,43.00247467252846],[-72.45131286188598,43.0000593327335],[-72.45591159592469,43.0012342223268],[-72.45972493649784,43.000473698147815],[-72.4630532363422,42.99686330843877],[-72.46453518764118,42.99405332021167],[-72.46557979742174,42.99001956156542],[-72.46430848141782,42.986363257989765],[-72.46201394395557,42.98383912603393],[-72.46215075401398,42.98227906273835],[-72.46585453640519,42.97838062916804],[-72.47049867107864,42.97539564709253],[-72.47336096426254,42.97230277303266],[-72.47590337575095,42.97147867457559],[-72.47931230251538,42.973456238775576],[-72.4818907317412,42.9738555495985],[-72.48723575744988,42.97162263915485],[-72.49213796651004,42.96773103666698],[-72.50277570824481,42.96620095587269],[-72.50663479226135,42.96444129231354],[-72.51473330257795,42.963790151232715],[-72.51849060933338,42.96310555380843],[-72.52498996247024,42.95860833353373],[-72.53227293043636,42.95479529574785],[-72.53456410570632,42.94986258297694],[-72.53183896536079,42.94669857690336],[-72.52817647937553,42.945088378422334],[-72.52692954427407,42.941345868034496],[-72.52634700858526,42.935484859791956],[-72.52706504648276,42.92817270504461],[-72.52429012553505,42.91866116536191],[-72.52457052938597,42.91538817828319],[-72.52604691938757,42.91384698867382],[-72.52946051658188,42.91247381394334],[-72.53154739891484,42.90723286124085],[-72.53098699213334,42.8990854253011],[-72.53175504528421,42.89665036029011],[-72.54056396287801,42.88944986525769],[-72.54728683599747,42.88675173041382],[-72.55060389398639,42.88624866102567],[-72.55263706235897,42.88498536521931],[-72.55239375094519,42.88345207622479],[-72.55519040541351,42.87499446232269],[-72.55539587254567,42.8667455464038],[-72.55412786230741,42.86080490542942],[-72.55719590522644,42.85412417196963],[-72.55705811465272,42.852773665143026],[-72.55415575316995,42.84903319484937],[-72.55245373313711,42.84572708330069],[-72.54908682626306,42.84283882585692],[-72.5471467156823,42.840020194516015],[-72.54674622416204,42.837099637388775],[-72.54717847691187,42.833210295934535],[-72.54653067024032,42.829401151750154],[-72.54666506563085,42.82596266122107],[-72.54549514678055,42.821933706788776],[-72.54235087492002,42.81589801638916],[-72.54263144799005,42.81172480623334],[-72.54172158924959,42.80746435167547],[-72.54080710899736,42.80658418424313],[-72.51836468584794,42.78961475577382],[-72.51233083293086,42.7844630952117],[-72.50896058710813,42.78006450623523],[-72.50879444909775,42.776207090917985],[-72.51033365225021,42.774600075657425],[-72.51383465767094,42.77330258368485],[-72.51624345441289,42.76998253806664],[-72.51652932301651,42.76842151893196],[-72.51495528865418,42.76514008158949],[-72.51070950408288,42.76382263686563],[-72.50804444605353,42.76439449382835],[-72.50573741860246,42.76601808626641],[-72.50118124543735,42.76732155963281],[-72.49907149102127,42.76972010209044],[-72.4986913236467,42.77182075649406],[-72.49674436386975,42.77356209761744],[-72.49125098797815,42.772728928184456],[-72.48696652984381,42.768714090618225],[-72.4855784429616,42.76636840661839],[-72.47990984338051,42.76348838543894],[-72.47660055229389,42.76010320558604],[-72.47510910247635,42.756193867830866],[-72.47467424585744,42.75130989876118],[-72.4736110533207,42.74668603476813],[-72.47268972906116,42.74531802262322],[-72.46768999780126,42.74150923389178],[-72.4667145464303,42.73942881131325],[-72.46332440035467,42.73655183601075],[-72.4590569846022,42.72985893965167],[-72.45844623396611,42.7269687026272],[-72.56535340629925,42.72996421398814],[-72.63975322956053,42.73199413510237],[-72.77129417101581,42.735586192844906],[-72.84635937936777,42.73737792595617],[-72.93010501964913,42.739199042091364],[-72.92981371666136,42.751353448758834],[-72.92829503458154,42.78737397274454],[-72.92659925159568,42.82337384665691],[-72.92740562375725,42.82981451847164],[-72.92627138296479,42.856364593043835],[-72.92502424045205,42.87984500794249],[-72.92406486393021,42.89541405276249],[-72.92333501824497,42.91516899696304],[-72.92379150493059,42.92745051470547],[-72.95383942154577,42.928491987661204],[-73.00671556786061,42.930755930396096],[-73.00813493033249,42.93056555007521],[-73.0076578214369,42.9400566468431],[-73.00510260521452,42.98509329020586],[-73.00312108739993,43.02369129812503],[-73.0005988754641,43.02357390307927],[-73.0006090994902,43.039129131755615],[-73.00178040086475,43.07513286277968],[-72.9983198002637,43.096606034567294],[-72.99869823473288,43.09749680805924],[-72.99714122011731,43.117466653741054],[-72.99596930625304,43.11870873956603],[-72.94841866218374,43.11749234335004],[-72.89827755566628,43.11465700220333],[-72.86824366379301,43.113939800661555],[-72.86731640840584,43.13185768575103],[-72.86480228958187,43.166157278383885],[-72.86342199588489,43.18367701590372],[-72.8621374913316,43.205224902351],[-72.85163044032134,43.204842260771976],[-72.85063656110718,43.21970898151562],[-72.84998238218628,43.23330694480721],[-72.84811353364823,43.26285536312888],[-72.81962056340674,43.25601438201351],[-72.75682319827382,43.239694705034054],[-72.70141356956857,43.226337546532754],[-72.68544184708287,43.22258898153981],[-72.67631750251499,43.22267352942886],[-72.67624428970716,43.22409200171442],[-72.60292208462033,43.23076809664846],[-72.5491533490645,43.235905583761415],[-72.4873383441494,43.23465704295029],[-72.4588539788752,43.2337682936334],[-72.43477237650221,43.23326480936767]]]},"properties":{"OBJECTID":5,"CNTY":25,"CNTYNAME":"WINDHAM","ShapeSTArea":2068043863.2131958,"ShapeSTLength":224823.94615743248}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-72.86864799426148,43.300537763988615],[-72.84956948656098,43.30018783120005],[-72.81989655729153,43.2993807446879],[-72.82003172808133,43.2918325660342],[-72.81962056340674,43.25601438201351],[-72.84811353364823,43.26285536312888],[-72.84998238218628,43.23330694480721],[-72.85063656110718,43.21970898151562],[-72.85163044032134,43.204842260771976],[-72.8621374913316,43.205224902351],[-72.86342199588489,43.18367701590372],[-72.86480228958187,43.166157278383885],[-72.86731640840584,43.13185768575103],[-72.86824366379301,43.113939800661555],[-72.89827755566628,43.11465700220333],[-72.94841866218374,43.11749234335004],[-72.99596930625304,43.11870873956603],[-72.99714122011731,43.117466653741054],[-72.99869823473288,43.09749680805924],[-72.9983198002637,43.096606034567294],[-73.00178040086475,43.07513286277968],[-73.0006090994902,43.039129131755615],[-73.0005988754641,43.02357390307927],[-73.00312108739993,43.02369129812503],[-73.00510260521452,42.98509329020586],[-73.0076578214369,42.9400566468431],[-73.00813493033249,42.93056555007521],[-73.00671556786061,42.930755930396096],[-72.95383942154577,42.928491987661204],[-72.92379150493059,42.92745051470547],[-72.92333501824497,42.91516899696304],[-72.92406486393021,42.89541405276249],[-72.92502424045205,42.87984500794249],[-72.92627138296479,42.856364593043835],[-72.92740562375725,42.82981451847164],[-72.92659925159568,42.82337384665691],[-72.92829503458154,42.78737397274454],[-72.92981371666136,42.751353448758834],[-72.93010501964913,42.739199042091364],[-73.04843976523922,42.741660988945085],[-73.16364059328392,42.74395326272433],[-73.26443808146271,42.74585271359056],[-73.27619343886758,42.74598938852248],[-73.27764428890295,42.750881122437875],[-73.28286376260645,42.77116506068292],[-73.29009170771631,42.79849570783469],[-73.29088950864312,42.8019640951329],[-73.28631517749513,42.808093138978485],[-73.28656580231845,42.80926942555357],[-73.28351965718615,42.81405869221046],[-73.28570365775413,42.81664369511463],[-73.28697021731462,42.82024994315005],[-73.28530942801754,42.834072243173544],[-73.28420274003956,42.83494258075418],[-73.27867914284607,42.83348108463789],[-73.27820616341147,42.83739954334213],[-73.27751352228636,42.857545838720974],[-73.2763277685559,42.89696295037928],[-73.27559223128488,42.91606303117954],[-73.27446259769732,42.94052235235812],[-73.27145947832541,43.00008854393463],[-73.27010413497746,43.02981103323968],[-73.26951488368039,43.037634774817],[-73.26730814133516,43.073652751879784],[-73.26673263520725,43.07977862141853],[-73.2654054110287,43.101033476174564],[-73.264277898585,43.12240482526779],[-73.26379350547313,43.12693044409409],[-73.26281397319148,43.15283998346127],[-73.25974447774468,43.20013951747281],[-73.25827633009304,43.23859340116057],[-73.25656578921549,43.2577107660559],[-73.25461006970804,43.31450665987397],[-73.2441024821414,43.31395126485172],[-73.21942654824204,43.31302862644884],[-73.16816682242377,43.31136774280403],[-73.13312771038603,43.30935743022102],[-73.09256990763464,43.30693898303623],[-73.04679880843494,43.30706793879609],[-72.99683647526814,43.30518272719097],[-72.98759097847974,43.30501359287805],[-72.99232803501846,43.290724407201495],[-72.97588639164515,43.29055941838742],[-72.97529615622253,43.30416746250894],[-72.94149213106009,43.30300197531988],[-72.89887414039559,43.30172192229425],[-72.86864799426148,43.300537763988615]]]},"properties":{"OBJECTID":6,"CNTY":3,"CNTYNAME":"BENNINGTON","ShapeSTArea":1753287168.5691814,"ShapeSTLength":215509.15083170377}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-72.55392229303604,45.00813729074689],[-72.55248533250622,45.00300225609495],[-72.55123785454646,44.99658426579739],[-72.53850917778888,44.935403935366374],[-72.53650006984168,44.92612408518761],[-72.54138462174261,44.92598662813111],[-72.53698928285543,44.8750039568626],[-72.53621528661253,44.86799211898105],[-72.53328173402815,44.834089982270065],[-72.57907591571808,44.780904784770726],[-72.594396767384,44.79030231109658],[-72.59789977567057,44.78706377735182],[-72.61673092139489,44.79585329297122],[-72.61359847507346,44.80042630799891],[-72.65170465128199,44.801705614133084],[-72.72909927907861,44.80404877632594],[-72.74685181919779,44.790098657136724],[-72.74579100255097,44.77591241456022],[-72.75504806746474,44.77704173897703],[-72.7585515828663,44.763649630719556],[-72.76859037089865,44.76500061307501],[-72.77439520924898,44.743179536285716],[-72.77938997197342,44.74384127562512],[-72.78241459393855,44.73306033915354],[-72.78882577621856,44.70914920333214],[-72.79322081613599,44.69067810331945],[-72.81706290816399,44.69926328551162],[-72.85079476563554,44.71113329845452],[-72.86368806183658,44.69360768122964],[-72.88798285988034,44.661005766311085],[-72.89186306267226,44.655669499505855],[-72.89439607282398,44.65476788046911],[-72.90081406612025,44.64977087088976],[-72.9029726984658,44.646364185736324],[-72.90560296018792,44.645915803539005],[-72.90989838190477,44.6466815088942],[-72.91630234081356,44.645813319773076],[-72.91879676041178,44.64680842141183],[-72.92264085255812,44.65012405534333],[-72.92666986950005,44.6475469116935],[-72.92728064756317,44.6457447872585],[-72.92375957569192,44.641951240630654],[-72.9265485157662,44.640328109364184],[-72.92396930411171,44.63741395749067],[-72.92641556108607,44.63513425664867],[-72.92377886089447,44.6332237262317],[-72.94253298456528,44.636115862088666],[-72.98352770589622,44.642584509790964],[-72.99432691552323,44.64436610715115],[-73.00929421259129,44.646511987647436],[-73.03783661824305,44.65129047227957],[-73.03807607862669,44.65115984590366],[-73.10282514170754,44.67616487782963],[-73.16068376809086,44.69751869943463],[-73.22885687298746,44.722817895851385],[-73.25013579223629,44.750224603789306],[-73.25529863079942,44.758081554348706],[-73.25797255202558,44.76359553580399],[-73.2588370531649,44.76809200039521],[-73.2588871386311,44.774096792880485],[-73.25591393442753,44.7831710278447],[-73.2526901120541,44.78901853156588],[-73.24574750976059,44.799588372202614],[-73.24291410052965,44.803326018684984],[-73.23703144649159,44.8096626806059],[-73.23006310953835,44.81469483351476],[-73.22412968111153,44.81815859501493],[-73.21687522591805,44.82062066681408],[-73.210821451737,44.82210130594112],[-73.20757657601354,44.82554374737249],[-73.20704826571769,44.8292256868831],[-73.20322210104487,44.8336399703634],[-73.20311722066053,44.83853502987461],[-73.20676654949014,44.843867745886854],[-73.20755552328397,44.846104393775],[-73.20705637918306,44.85195675950775],[-73.20351034473015,44.861803716947705],[-73.20333236769989,44.865995492476664],[-73.19830643782595,44.871848503069764],[-73.19643351387317,44.87542888221953],[-73.19567185557246,44.882978239113385],[-73.19304145672403,44.88894164653275],[-73.19199539609963,44.89252583022596],[-73.19144757967253,44.89807970533099],[-73.19131868889878,44.90561221238166],[-73.19231092187229,44.908302150006776],[-73.19534916916606,44.910591070524134],[-73.20569845372954,44.91110575813656],[-73.21164889382469,44.913098058373386],[-73.21594625228322,44.91526321129958],[-73.22377659965154,44.91785139428439],[-73.22612632272815,44.91960725686122],[-73.23244529070674,44.93066040061028],[-73.23425103956849,44.93701713089356],[-73.23795114352099,44.94064758060852],[-73.24041127052149,44.945276984819074],[-73.24094565413421,44.94889469800273],[-73.23913032448247,44.9532238696633],[-73.22896608209605,44.959988746843905],[-73.22326801487539,44.96703856753372],[-73.22010863695098,44.97258913132462],[-73.21629732083638,44.97775944966089],[-73.20699129309702,44.98961389184801],[-73.20106588489008,44.99827032524488],[-73.19635850423604,45.0057668319345],[-73.19231232645926,45.013479954714725],[-73.09026265133551,45.01534510702272],[-73.06536303317633,45.01597074178568],[-72.99598664810102,45.01478306022607],[-72.96685381971986,45.01415657352297],[-72.90865656942727,45.01560692063168],[-72.86907346057868,45.016212749096105],[-72.84700691671165,45.01667401293214],[-72.82132376736399,45.01629087689432],[-72.74817711034669,45.01563036897126],[-72.73390719740138,45.01583896771879],[-72.70293775132733,45.01562755927616],[-72.63467728312364,45.01469697871297],[-72.60430285748122,45.013121982162005],[-72.58475143249659,45.01157754349645],[-72.55539937869345,45.00802759497162],[-72.55392229303604,45.00813729074689]]]},"properties":{"OBJECTID":7,"CNTY":11,"CNTYNAME":"FRANKLIN","ShapeSTArea":1784780014.1212387,"ShapeSTLength":188538.47249967672}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-71.83790101385733,44.34797859500394],[-71.86044903445845,44.36996294862],[-71.87839931250521,44.38782214586725],[-71.90605067879153,44.41459777379239],[-71.93322334951141,44.44114742619373],[-71.90326477238696,44.45475140036407],[-71.88074994306253,44.47471741227188],[-71.86825437184466,44.4860422975662],[-71.85669166095646,44.49628382879252],[-71.86087066757801,44.50075651986754],[-71.87458673872642,44.51460929971547],[-71.90893941532029,44.54792625928389],[-71.90067800267465,44.55547433832796],[-71.89080781045557,44.565437929181314],[-71.86800239750346,44.586560303221034],[-71.84237299587912,44.610876171649565],[-71.86121676541812,44.62114849368872],[-71.87194396692443,44.62668654419641],[-71.89581218123307,44.63938326177597],[-71.91019935216197,44.646840337316796],[-71.87018634997737,44.68452264817609],[-71.83706847897619,44.71638511143216],[-71.85503532369306,44.72563268143116],[-71.87456849801585,44.73590901251434],[-71.90224460673875,44.75004040022158],[-71.93500842573079,44.76729175118564],[-71.93929273719223,44.76945013143421],[-71.97482609720078,44.78724062557453],[-71.93683910186412,44.822541420493785],[-71.92931744591431,44.82964832537888],[-71.91995085333807,44.83690375089462],[-71.89809815971485,44.85887096608518],[-71.87785847130161,44.87877145078154],[-71.88768224564625,44.88414402607262],[-71.92014649911405,44.901534959428105],[-71.91256962807861,44.941586303684424],[-71.90639495793083,44.972034737155035],[-71.89930556685702,45.007967038225665],[-71.8290277583976,45.009743466634575],[-71.80495524609474,45.01046019264999],[-71.75410185388803,45.011346436014875],[-71.6976845037793,45.01141116023102],[-71.69679749293314,45.011307710152394],[-71.66672995007666,45.01170501805547],[-71.63956013222214,45.01237719975212],[-71.61265699150408,45.01272438193821],[-71.57251131586001,45.01250767774122],[-71.50102024854127,45.013454940303305],[-71.49260709230421,45.0134278674576],[-71.46547392797625,45.013680240691826],[-71.46599559775555,45.01204412337721],[-71.47298489537702,45.01104418676312],[-71.47845017514375,45.00879513840556],[-71.4801436812111,45.00301074408847],[-71.48109793243306,45.002378828043874],[-71.48691940335543,45.00228592337139],[-71.48822161854142,45.00091122732796],[-71.49295703513646,45.002805456210574],[-71.49571264872759,45.004901120828926],[-71.49852552083937,45.00390438732051],[-71.5011875636768,45.00695385597818],[-71.5044578393425,45.00839600826404],[-71.50917200187317,45.007925705954364],[-71.51453199657115,45.00443759984302],[-71.52160232400998,45.00226925255524],[-71.52412686249218,45.00242844431515],[-71.53180834218307,44.998653475536635],[-71.53609393232657,44.994832637408045],[-71.53708584466679,44.992505405099806],[-71.53369170304376,44.9917351616716],[-71.53562023854163,44.98937446152318],[-71.53826240730943,44.9896568052898],[-71.53879158825333,44.98720606314533],[-71.54146046159572,44.98525715891453],[-71.53673877528593,44.983774942333355],[-71.53755229546846,44.98131274539938],[-71.53247431389563,44.97936697385609],[-71.53099971378934,44.974899762138],[-71.5273687798799,44.97301174620513],[-71.52653110924439,44.97057370132703],[-71.52409538722831,44.96726271902117],[-71.52207561610815,44.965955297881514],[-71.52319906469322,44.963595555951606],[-71.51928461993842,44.96320583606276],[-71.5164437601857,44.964235420473464],[-71.5152641264238,44.95849730852401],[-71.51605556261313,44.95277617828052],[-71.51954365315596,44.949932347706536],[-71.51707203265407,44.948389775667515],[-71.51876461102195,44.944716112683544],[-71.51530056279931,44.94398394384953],[-71.5158216411757,44.9399766192787],[-71.51680800850775,44.93894459158462],[-71.52123583745173,44.937789844967114],[-71.51624089148896,44.934979128937506],[-71.51596351263206,44.93127794329882],[-71.51328533175189,44.92893342671302],[-71.5158411262767,44.926665027821585],[-71.5126693309352,44.92615189366819],[-71.50985974136003,44.92406811158549],[-71.50722530857186,44.92077616725688],[-71.50456375147422,44.918881612352536],[-71.50123378384893,44.91472074992792],[-71.49973190741139,44.91074184159817],[-71.49532624554084,44.91164983801813],[-71.49668050235431,44.90809147741741],[-71.49500066612138,44.90470296843117],[-71.49672323443076,44.903706200049704],[-71.50104251971491,44.9046273886891],[-71.50379513452991,44.901403860358236],[-71.50762335369284,44.89969673375051],[-71.50849567408403,44.896479590683505],[-71.5142382406675,44.894670241879595],[-71.51344503122378,44.89188959841265],[-71.51098756564579,44.890548759798754],[-71.51560020483512,44.88709436995081],[-71.51701011162915,44.88465604575708],[-71.51875205986661,44.885149314617735],[-71.52123789210455,44.882973944373596],[-71.5222278115792,44.880487228447734],[-71.5248137292149,44.87890062019772],[-71.5289003533277,44.877798982541435],[-71.52941963150617,44.8734033251642],[-71.53313048955619,44.87035557908705],[-71.53741308493822,44.868740110934134],[-71.54036336425962,44.8687077988322],[-71.54622405659548,44.86618304601517],[-71.54996671046044,44.86214280202085],[-71.55059608056196,44.8605099735929],[-71.54924278903967,44.85837876317312],[-71.54727143805734,44.85799652767743],[-71.54799515575618,44.85511617727657],[-71.55081670923106,44.85416679657998],[-71.55598579170675,44.84974989263003],[-71.55599065962504,44.846788041193705],[-71.55231127107618,44.84277654875714],[-71.55123809359671,44.83746287907455],[-71.55356153720341,44.83541790153128],[-71.55739500514589,44.834347151931304],[-71.55940424362102,44.83076136905167],[-71.56152840683767,44.82482354399984],[-71.56466016301577,44.824392092913826],[-71.57391855950348,44.82060431263687],[-71.5770826957398,44.81561086640258],[-71.57662373842473,44.81250711348895],[-71.57317910411952,44.81101013140835],[-71.56870045297163,44.806637915590194],[-71.5705846845015,44.80281278516496],[-71.57323734100142,44.79943192903109],[-71.57166646684841,44.79354050767205],[-71.57327338893414,44.790910654370954],[-71.57593862895779,44.789634599569766],[-71.57936051801421,44.78512648401924],[-71.58387267358445,44.78523512190986],[-71.59266275167838,44.78278297676124],[-71.59658765509991,44.77871397593687],[-71.59524264893615,44.77642712305136],[-71.59632501688634,44.77505038813043],[-71.60163867324688,44.772763642307204],[-71.60339242528829,44.769303090740614],[-71.60635769338417,44.767287517959],[-71.61169584457103,44.765472760612376],[-71.61371294947521,44.76104148063761],[-71.61322528399644,44.7590756389263],[-71.61667171951545,44.75691414797239],[-71.61868736306829,44.75708684412542],[-71.62382019224903,44.75534894237716],[-71.6253352399117,44.75410159342463],[-71.63111857841157,44.753319271474886],[-71.63281233572492,44.75103541405958],[-71.63215050001999,44.748636641438914],[-71.62742333851659,44.746995643343745],[-71.62564860482651,44.74337849205143],[-71.62687183587454,44.740397107847656],[-71.62535153783168,44.73737952099235],[-71.62678645447248,44.7318005571496],[-71.6266895490341,44.7286263859732],[-71.62352223741931,44.72692109494424],[-71.61759220113272,44.72874255393792],[-71.61919368757975,44.72207814674338],[-71.6171178232724,44.72103163262868],[-71.61469537093232,44.717762859839155],[-71.60896671721643,44.71370246470693],[-71.60605017521826,44.70917283557646],[-71.60420943560794,44.70749879790425],[-71.6013399165398,44.70651433847494],[-71.59939933655818,44.70440856155685],[-71.59939876565669,44.70250848507902],[-71.60103533723641,44.7011713344146],[-71.60142914683715,44.69880532788115],[-71.59848915787948,44.69722971814061],[-71.59478505111448,44.69671230710566],[-71.5988369779471,44.69223705258109],[-71.59733945276497,44.68891651735383],[-71.5988840655023,44.68765479918106],[-71.59911039352227,44.685106502723315],[-71.59543056175649,44.681992418861604],[-71.59731469325347,44.67891207380524],[-71.59626975148576,44.67743235843509],[-71.59301251747978,44.67616366454639],[-71.59010015486518,44.6763913934503],[-71.58811880318304,44.67411941320505],[-71.58245503266808,44.67401483268047],[-71.58207523430714,44.67214085784346],[-71.58783702839352,44.670519210685285],[-71.58791198364268,44.66877163401172],[-71.58551575299224,44.66628076059985],[-71.58564590888747,44.66288650880154],[-71.58837828751004,44.660584227590086],[-71.58542745275538,44.65694737583237],[-71.58039623558379,44.65501057598154],[-71.57639619725423,44.65608780427927],[-71.5748854855473,44.65396364897732],[-71.57998655760737,44.651385735446866],[-71.57586599415099,44.64998566912422],[-71.5712179999405,44.64945750734908],[-71.56840156768708,44.65016556907018],[-71.56584462863552,44.65381149030599],[-71.56326052014016,44.65249974965074],[-71.5618889609847,44.646361665289916],[-71.55946067897256,44.644473819581265],[-71.55833248647745,44.64085836393811],[-71.56047386607447,44.63939257266384],[-71.56394474558857,44.63922822174301],[-71.5637992185511,44.63666981472302],[-71.55846789191037,44.63413059439453],[-71.55458520291543,44.63030318920539],[-71.55159586744129,44.629585463158726],[-71.54969708937848,44.630495968906914],[-71.54701202203277,44.62892347800767],[-71.55117178327546,44.62778425653182],[-71.55424368324344,44.62953452805229],[-71.55709971930848,44.62943994523432],[-71.5565344256089,44.627191207699866],[-71.55264510616865,44.62652346892447],[-71.5514964160518,44.62466548287969],[-71.54821856577314,44.625267797273814],[-71.54636993186894,44.62305962492425],[-71.54767226699548,44.62196331570815],[-71.55218084948405,44.624638587319495],[-71.55555548905907,44.62509535937347],[-71.55679138756565,44.61776742124556],[-71.55359873800995,44.60707042447488],[-71.55412287807913,44.60520220215582],[-71.55726193896757,44.60324715272085],[-71.5618789526896,44.60333466774112],[-71.56276011650843,44.60222041501988],[-71.56113966049779,44.59977604502667],[-71.55604553431054,44.59927688814031],[-71.55134569940643,44.60116423655011],[-71.55027325561129,44.59969573349978],[-71.55449695219075,44.597975922821156],[-71.55419725975312,44.59320373817773],[-71.55226689022139,44.59339094034294],[-71.54778214087186,44.5972764165276],[-71.5455483895321,44.59541273968764],[-71.54768834199525,44.5918764307472],[-71.54471384495321,44.59023171943019],[-71.53953131641175,44.59403498987767],[-71.53759356919478,44.5929324652388],[-71.54035822764226,44.58989813437823],[-71.53650720801356,44.58682512255731],[-71.53824363242553,44.58234434562281],[-71.54478705869002,44.578479982306185],[-71.54800050895385,44.577744843544366],[-71.55142097188538,44.58034917449396],[-71.55047487448343,44.58341762891732],[-71.55321014790914,44.58383081821148],[-71.55507652511541,44.58029848468392],[-71.55294214757521,44.57612349063309],[-71.54835577569325,44.57259948523449],[-71.54907216506345,44.569724324347085],[-71.55169398968535,44.568998552426244],[-71.55788595946133,44.57061504941143],[-71.55924924649888,44.56856355660812],[-71.55843568361357,44.56504775025124],[-71.56218667827113,44.563251121465065],[-71.57079240736553,44.56296131218665],[-71.57587042590839,44.56513652440894],[-71.58149214237271,44.564800934267126],[-71.58812731016204,44.56593753459589],[-71.59034505383808,44.56575876473911],[-71.59348053631389,44.564125843493144],[-71.59731870292292,44.55930479773111],[-71.59833162043209,44.55490186904574],[-71.59774282211812,44.55394184945905],[-71.58690889292575,44.5469505490786],[-71.58409934084983,44.54612836886645],[-71.57798167205151,44.5429678894506],[-71.57396336182182,44.540166822685514],[-71.57181397720379,44.537200367243294],[-71.57336517697578,44.53358774932139],[-71.57987718183612,44.52562781057561],[-71.58331975487194,44.523167523685196],[-71.58892841396606,44.522334335775724],[-71.59301778333024,44.5231526202922],[-71.59405919428517,44.520386914527556],[-71.59110326479832,44.51731569610264],[-71.58627705996258,44.51469966834655],[-71.58373995539063,44.50910304536831],[-71.57823097765878,44.50535563920578],[-71.5768342635464,44.50241261928559],[-71.5788712878079,44.501405203344156],[-71.58246734284741,44.50201749147305],[-71.58416544161202,44.50327052052637],[-71.58617231134423,44.506786916856235],[-71.58870001211415,44.50606304304314],[-71.58699046238786,44.503848493252725],[-71.58568763835635,44.49963498347974],[-71.58953210025766,44.49833423112691],[-71.59054059656076,44.50087812370643],[-71.59375299342855,44.501293604154135],[-71.59459930199728,44.50008890486736],[-71.59491919138007,44.49423704458334],[-71.59716092831907,44.49061424906623],[-71.59825767307288,44.48742019351535],[-71.60121186004957,44.48554720997241],[-71.61075243469814,44.48446482696052],[-71.61555541500839,44.48612949799872],[-71.61881976798973,44.485264775828554],[-71.62335636664483,44.478010495117566],[-71.62438281226991,44.47563144253185],[-71.62740604253331,44.47428360312371],[-71.63064905201124,44.47402134161134],[-71.63402414820746,44.47515148702736],[-71.63511845422235,44.4776329181977],[-71.6273914336288,44.479868373887825],[-71.62443691881849,44.48200661996028],[-71.62564847654048,44.48384036040762],[-71.63228323464145,44.48416481243019],[-71.63350620417026,44.481740017890424],[-71.63659312819536,44.48047008428095],[-71.63886451925144,44.477874962561614],[-71.64258146438917,44.47694052562451],[-71.64775745210704,44.473199223074126],[-71.64768604333293,44.47010926375455],[-71.64509193860937,44.46814329030684],[-71.6404331929569,44.466120070592055],[-71.63955457684722,44.46425020245485],[-71.6430133361805,44.4613534460701],[-71.64673426538594,44.46040504345398],[-71.65236473412567,44.46103328477866],[-71.65557451965076,44.45805690243529],[-71.65705563789253,44.455049775697056],[-71.65779084241983,44.44813206659081],[-71.65915620292405,44.44401720333948],[-71.66220224610323,44.43978587853935],[-71.66441950434638,44.438126411431774],[-71.67006156385659,44.43633407697679],[-71.67775482912442,44.436059349732986],[-71.67954062378973,44.43461075766583],[-71.67665658710389,44.4290687939412],[-71.67779173256434,44.42720182836031],[-71.68037376676142,44.428202595991976],[-71.68369905609474,44.4271340469686],[-71.68486531674377,44.42391298197322],[-71.69024000874289,44.421440432259175],[-71.69413166744376,44.4202641776677],[-71.69658231186216,44.41811164934585],[-71.70231079779157,44.414633696606025],[-71.7088677991889,44.41186954971594],[-71.71442428300925,44.410279791416244],[-71.71894239910411,44.4105607853698],[-71.7222887011889,44.41153901329592],[-71.72597904427936,44.411629599325515],[-71.73602924481541,44.4102626448118],[-71.73961636587026,44.40709434455351],[-71.7428015184017,44.40154036043428],[-71.74727174343488,44.40116771599047],[-71.74951297390047,44.401944457828186],[-71.75402441436827,44.40529170738171],[-71.75684800747972,44.406597771814965],[-71.76259391936195,44.407072392956714],[-71.76785434113694,44.40551348875192],[-71.77274813217542,44.403126567621705],[-71.7778939474081,44.40002673338504],[-71.78053249155313,44.39973649844591],[-71.79213782396644,44.4001052768778],[-71.79542731812262,44.39844296366607],[-71.80301274257933,44.39274236591174],[-71.8037249342903,44.39077120729485],[-71.79993321715656,44.38611278635509],[-71.80038897386181,44.384189482122686],[-71.80337086711044,44.383290558448486],[-71.8091317255164,44.38376445882734],[-71.81380236947037,44.38251962011991],[-71.81469790255035,44.38131301799559],[-71.8156376354557,44.37496186200808],[-71.81266383582992,44.37270039903808],[-71.81274487630485,44.370824195135086],[-71.8155209504931,44.36892889221591],[-71.81627708874873,44.36734597551222],[-71.8149575760555,44.36326015325405],[-71.81242015931092,44.35756891942927],[-71.81411274816625,44.35484473315694],[-71.81837786292243,44.35354443384498],[-71.82578052149714,44.35296271407736],[-71.82885242534464,44.35116917470198],[-71.83319273492297,44.35024207075615],[-71.83790101385733,44.34797859500394]]]},"properties":{"OBJECTID":8,"CNTY":9,"CNTYNAME":"ESSEX","ShapeSTArea":1740875031.664795,"ShapeSTLength":275955.1615190162}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-72.57907591571808,44.780904784770726],[-72.52019325610684,44.75974616084036],[-72.50087300480874,44.752627096688336],[-72.4224471271231,44.72447131979226],[-72.41767326776278,44.722688949026555],[-72.45420358363344,44.67313270642213],[-72.48279415530956,44.634112332199315],[-72.45815134995883,44.62505021122964],[-72.37112518348113,44.593344555644464],[-72.36838584940472,44.59285547005072],[-72.37175984409885,44.58823645937912],[-72.37386153254992,44.584171294492144],[-72.39856764023166,44.551815318622985],[-72.40744334889284,44.54029592616762],[-72.4284780384703,44.512566878811],[-72.42981185855568,44.51202684102386],[-72.43468432989302,44.505247853204025],[-72.44734333991018,44.48765525380602],[-72.46699826029956,44.46154010093318],[-72.47447865178309,44.45225470415154],[-72.47593437169152,44.45001944988791],[-72.4847430738684,44.439144316181896],[-72.48509149867076,44.436778383469836],[-72.49324310211328,44.425837594454464],[-72.49356241633227,44.4239855935797],[-72.51236044843273,44.433913952477916],[-72.5595582810994,44.45908075347308],[-72.59728607324982,44.47939396396597],[-72.60025080674642,44.47402420205444],[-72.60524465740579,44.46731608203702],[-72.62468188578923,44.44010646811361],[-72.65434084524145,44.39830648779421],[-72.68072028336027,44.409207250476335],[-72.70582918643258,44.419462616837194],[-72.73434433301003,44.4308277503959],[-72.76556049213833,44.442814590279625],[-72.78606661677529,44.44741755689836],[-72.80320710340565,44.4515249990348],[-72.8382972186703,44.45930183682008],[-72.80362994552533,44.5062898887272],[-72.83079748289437,44.51643552620223],[-72.80995811259517,44.54672310649224],[-72.82704363164359,44.55070069488409],[-72.85846799278407,44.579787270846],[-72.87919405921666,44.60071893243115],[-72.90996235025344,44.63120631512025],[-72.92377886089447,44.6332237262317],[-72.92641556108607,44.63513425664867],[-72.92396930411171,44.63741395749067],[-72.9265485157662,44.640328109364184],[-72.92375957569192,44.641951240630654],[-72.92728064756317,44.6457447872585],[-72.92666986950005,44.6475469116935],[-72.92264085255812,44.65012405534333],[-72.91879676041178,44.64680842141183],[-72.91630234081356,44.645813319773076],[-72.90989838190477,44.6466815088942],[-72.90560296018792,44.645915803539005],[-72.9029726984658,44.646364185736324],[-72.90081406612025,44.64977087088976],[-72.89439607282398,44.65476788046911],[-72.89186306267226,44.655669499505855],[-72.88798285988034,44.661005766311085],[-72.86368806183658,44.69360768122964],[-72.85079476563554,44.71113329845452],[-72.81706290816399,44.69926328551162],[-72.79322081613599,44.69067810331945],[-72.78882577621856,44.70914920333214],[-72.78241459393855,44.73306033915354],[-72.77938997197342,44.74384127562512],[-72.77439520924898,44.743179536285716],[-72.76859037089865,44.76500061307501],[-72.7585515828663,44.763649630719556],[-72.75504806746474,44.77704173897703],[-72.74579100255097,44.77591241456022],[-72.74685181919779,44.790098657136724],[-72.72909927907861,44.80404877632594],[-72.65170465128199,44.801705614133084],[-72.61359847507346,44.80042630799891],[-72.61673092139489,44.79585329297122],[-72.59789977567057,44.78706377735182],[-72.594396767384,44.79030231109658],[-72.57907591571808,44.780904784770726]]]},"properties":{"OBJECTID":9,"CNTY":15,"CNTYNAME":"LAMOILLE","ShapeSTArea":1207382477.3441925,"ShapeSTLength":166207.24754977846}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-71.83790101385733,44.34797859500394],[-71.84209934053733,44.3451911762607],[-71.85286896878831,44.340850508307895],[-71.86043442234883,44.34042279635714],[-71.86432546210251,44.33959323716493],[-71.87171279279963,44.33670498303584],[-71.87540668217734,44.33770674702741],[-71.88133127157414,44.34010500625517],[-71.88451786093671,44.34194978695577],[-71.89165477463163,44.34332460912259],[-71.90039239394576,44.34721929301715],[-71.90484132814693,44.348183386061336],[-71.90882709821074,44.34813721663737],[-71.91718469003018,44.34672518841036],[-71.92466238422918,44.342297909628655],[-71.92992317859526,44.337340802124906],[-71.93458120598707,44.33592337532462],[-71.93842712425032,44.3357918722669],[-71.9446758755577,44.33766061324359],[-71.9579318030374,44.33750406286563],[-71.9619384999499,44.33661415160841],[-71.97260128917108,44.336744943985195],[-71.97741177793472,44.33750309572013],[-71.9823667010368,44.33717840334249],[-71.98464200250082,44.336161556072575],[-71.98655868983623,44.33142770860551],[-71.9887540472281,44.329851448479204],[-72.00270726378699,44.324529243137754],[-72.00948653523797,44.322651157882675],[-72.01593211269562,44.32238067309506],[-72.02230139934822,44.321107478186924],[-72.02608523813896,44.322142817163105],[-72.02977279732684,44.32217452627984],[-72.03279624154807,44.32049309603526],[-72.03363945265754,44.31733527330678],[-72.03149502310137,44.31543433416774],[-72.03263705465956,44.30990939797725],[-72.03221988366342,44.30675922274022],[-72.03277910973239,44.30236666748806],[-72.03762672516743,44.29707303010987],[-72.04027170152004,44.29574850111859],[-72.04432086695736,44.29518262218923],[-72.04649636809017,44.29204811379756],[-72.05350734375098,44.290430036880984],[-72.05662146749685,44.287462160051525],[-72.05942424964385,44.285963475577184],[-72.06090359520626,44.283159351224725],[-72.06455151594487,44.27900326908723],[-72.06611057369277,44.27545527514413],[-72.06827010675224,44.272481017412595],[-72.06832034124118,44.270135678787035],[-72.06705494868227,44.26815156418265],[-72.0639445322548,44.268083125437535],[-72.06211087647654,44.26957034783209],[-72.05859979100302,44.26989315812282],[-72.05963456774688,44.2677867881036],[-72.058636631373,44.26497013834302],[-72.06056730163077,44.26435551592052],[-72.0594174628349,44.261874609763126],[-72.05965454882504,44.25660379103435],[-72.05581222681614,44.250558584718185],[-72.0548544823069,44.24798603525513],[-72.05020030044479,44.24429380267653],[-72.04801402412717,44.23983925018497],[-72.05363657775322,44.225684516705776],[-72.05391158153024,44.22269709441681],[-72.05244456220896,44.21911066645274],[-72.05278976568819,44.217251652596765],[-72.05536383696244,44.213036886135484],[-72.05874551495235,44.21169677602674],[-72.0598763820048,44.20988846880863],[-72.05773328398504,44.20605361771794],[-72.05896971192679,44.202277540410925],[-72.06060970950558,44.19986090452998],[-72.06444974049141,44.19757181283464],[-72.06439153444748,44.19545553588456],[-72.06612011585906,44.1909493697242],[-72.06559719153525,44.1889323014578],[-72.06071288140174,44.18430593005681],[-72.05802353943352,44.18051531449705],[-72.05651566011693,44.17586010110859],[-72.05464638926608,44.17253400863139],[-72.05332407241504,44.16855212585591],[-72.05158386845798,44.16626535006073],[-72.04856891538016,44.16469666291844],[-72.04700552978363,44.161496657869684],[-72.042151809732,44.16141151906999],[-72.04396272042044,44.158478741289386],[-72.04363993445739,44.15709928646535],[-72.15528201692726,44.187708317648294],[-72.17069198522198,44.19198122271944],[-72.17436575583575,44.192669627742355],[-72.19992598123842,44.19077116196788],[-72.25441585486044,44.18684745457586],[-72.30520205476473,44.18309561650894],[-72.34726079172276,44.19694973182391],[-72.36888140280671,44.20389734569612],[-72.35060906355639,44.2282211421443],[-72.33687028611274,44.24622356994914],[-72.33509504348709,44.24752419796214],[-72.31665271858014,44.27223731016606],[-72.31683527392694,44.298790099871],[-72.27688092354514,44.33424160917313],[-72.27009732807778,44.340381969881975],[-72.24406847917456,44.37504605920377],[-72.2170795103801,44.41131975338504],[-72.22204270867296,44.4142496015238],[-72.22011650170046,44.421170785733445],[-72.22193291865825,44.422844640791496],[-72.22521981200128,44.42169084186523],[-72.22720439766418,44.41939882013177],[-72.22892643882552,44.421717517721575],[-72.22314295816591,44.4238097230508],[-72.22380692971156,44.42489739713071],[-72.26316084321657,44.44031416216949],[-72.31743310412898,44.46110216260642],[-72.3213952644,44.46317534190703],[-72.3259063133564,44.46450883247986],[-72.38821084532559,44.4883454231688],[-72.42625290392988,44.50298755253065],[-72.43468432989302,44.505247853204025],[-72.42981185855568,44.51202684102386],[-72.4284780384703,44.512566878811],[-72.40744334889284,44.54029592616762],[-72.39856764023166,44.551815318622985],[-72.37386153254992,44.584171294492144],[-72.36653172204848,44.58079434275957],[-72.35419116580978,44.57644905463246],[-72.30319817668146,44.55778858931625],[-72.28771378786477,44.55181867632704],[-72.26289163479404,44.54250346778505],[-72.2616720312797,44.54376984071134],[-72.22889633382701,44.58798400220501],[-72.20712445792117,44.6167421524504],[-72.2024891506249,44.62410595577774],[-72.19926218787573,44.628291149617525],[-72.18495758607976,44.64925658454629],[-72.14996003178632,44.6978650872916],[-72.1459577724481,44.704507419885516],[-72.12452275235108,44.69729842843127],[-72.08929717635908,44.68509008796579],[-72.10194093000278,44.71773911466352],[-72.11197436534691,44.74320799587801],[-72.11436109432913,44.74982120146015],[-72.08972446262909,44.73767356435365],[-72.05216258622916,44.718994083231216],[-72.01145503900082,44.698841717258844],[-71.9995389566063,44.711086418730154],[-71.99123898495328,44.71882667502445],[-71.9750836685309,44.734647555007385],[-71.95140936345551,44.757495286499],[-71.93929273719223,44.76945013143421],[-71.93500842573079,44.76729175118564],[-71.90224460673875,44.75004040022158],[-71.87456849801585,44.73590901251434],[-71.85503532369306,44.72563268143116],[-71.83706847897619,44.71638511143216],[-71.87018634997737,44.68452264817609],[-71.91019935216197,44.646840337316796],[-71.89581218123307,44.63938326177597],[-71.87194396692443,44.62668654419641],[-71.86121676541812,44.62114849368872],[-71.84237299587912,44.610876171649565],[-71.86800239750346,44.586560303221034],[-71.89080781045557,44.565437929181314],[-71.90067800267465,44.55547433832796],[-71.90893941532029,44.54792625928389],[-71.87458673872642,44.51460929971547],[-71.86087066757801,44.50075651986754],[-71.85669166095646,44.49628382879252],[-71.86825437184466,44.4860422975662],[-71.88074994306253,44.47471741227188],[-71.90326477238696,44.45475140036407],[-71.93322334951141,44.44114742619373],[-71.90605067879153,44.41459777379239],[-71.87839931250521,44.38782214586725],[-71.86044903445845,44.36996294862],[-71.83790101385733,44.34797859500394]]]},"properties":{"OBJECTID":10,"CNTY":5,"CNTYNAME":"CALEDONIA","ShapeSTArea":1703495432.5658798,"ShapeSTLength":255199.55614223937}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-72.04363993445739,44.15709928646535],[-72.0420284779774,44.156635190118884],[-72.03960087374963,44.15600677259608],[-72.04308233599538,44.15099687292914],[-72.04225044879105,44.148442586843565],[-72.04159854290319,44.13988961832974],[-72.04193422729601,44.13675410262023],[-72.03830681314604,44.13380576827461],[-72.03326743861767,44.131923245198266],[-72.03579789364748,44.12918120440364],[-72.03742262707645,44.12579325674765],[-72.04062046001262,44.12614059978489],[-72.04584133068688,44.124536879912675],[-72.04935258646707,44.11974593934108],[-72.0533628464588,44.11883567300734],[-72.05340289948992,44.11449664124677],[-72.05540144215676,44.111461700364096],[-72.05284970371339,44.1040962701518],[-72.05299130323175,44.101863730388274],[-72.05006986867144,44.097825273013434],[-72.04740000392917,44.0964171960017],[-72.0447204699546,44.096269568559705],[-72.04286116313769,44.09742306305815],[-72.0416438303882,44.10181075805673],[-72.03805808162883,44.10344233609612],[-72.03264546611769,44.10150043128863],[-72.03070165703953,44.097938518114205],[-72.03229558029739,44.09190546643003],[-72.03385049635564,44.0900663552817],[-72.0381347768199,44.08869657274141],[-72.04325746121341,44.08899814920963],[-72.04506003447806,44.08659322891873],[-72.04217492729909,44.084729216405144],[-72.03743763204443,44.0827039241949],[-72.03384434369987,44.081786972392315],[-72.03162132973947,44.079837549612755],[-72.0341614209111,44.077115415056554],[-72.03843741927422,44.07680237498184],[-72.04150302422323,44.07921981475355],[-72.04434723676006,44.07931453508996],[-72.04715616142063,44.07811176982481],[-72.05169188035337,44.07728076710892],[-72.05248538206577,44.07445019577908],[-72.04875933567784,44.071423205706424],[-72.04880249449178,44.06901233645508],[-72.05415242974749,44.06384810035316],[-72.05668564681783,44.05905666267895],[-72.05855946930157,44.05796851209226],[-72.06547541718534,44.058391298512035],[-72.06827102922338,44.05738209154024],[-72.06907847692221,44.05418532287296],[-72.06747290067052,44.05306296805048],[-72.06299770251334,44.05198743258843],[-72.06201721664347,44.049616100884656],[-72.06649077267275,44.04907659224941],[-72.07628338525255,44.04508018130554],[-72.07908793232983,44.04246538356311],[-72.07944346504534,44.039335357712105],[-72.07491436142156,44.032877179595886],[-72.07686210599546,44.03042349129808],[-72.08029950971249,44.02948548099071],[-72.08256103317092,44.026460525568254],[-72.08198522185381,44.02385385652223],[-72.08266713960968,44.021915640818364],[-72.08646117790678,44.02147996528643],[-72.09114830361524,44.02433351223606],[-72.09435987229647,44.02343117019511],[-72.09539654940997,44.022032797582284],[-72.09592782126887,44.01848158118323],[-72.09476282985491,44.01534466688853],[-72.09044378301351,44.01121851488637],[-72.09334953836911,44.00968994711398],[-72.09371564259426,44.01345666613017],[-72.09813752337939,44.015451627924946],[-72.10046173514384,44.01548623932975],[-72.10429803459948,44.014203109751],[-72.10602069035266,44.01158514421785],[-72.10371316677436,44.00320390811725],[-72.10497005504787,44.00199890000002],[-72.10980871870704,43.99997436645071],[-72.11279550590609,43.99661799037375],[-72.11688938671699,43.994386143302684],[-72.11684161203281,43.99191010939729],[-72.11395019474747,43.989530974215384],[-72.1122806495357,43.986974237689964],[-72.11184749225764,43.984487313047445],[-72.11314380771972,43.981020942421466],[-72.11263682171972,43.97848012396789],[-72.11349843432956,43.972736683720946],[-72.11481488782202,43.96930005503199],[-72.11411320717222,43.96755629929254],[-72.11118691173992,43.966942204483104],[-72.10721679613799,43.96963437959142],[-72.10507911762993,43.9700407547965],[-72.09232717039578,43.96687213597465],[-72.09122045972504,43.96551807140617],[-72.09818651412135,43.964122985671054],[-72.10095574526709,43.96292116129384],[-72.10105843836027,43.96051637405656],[-72.09930705474201,43.959292777255705],[-72.09969906565804,43.95674616926919],[-72.10475337808609,43.950792517082064],[-72.1070714983722,43.94915741567789],[-72.11214166579778,43.947525612645045],[-72.11819849122033,43.946813835565344],[-72.11911164362698,43.943822298854414],[-72.11872862393574,43.94056452747447],[-72.11690240337788,43.934939708355074],[-72.1180310377826,43.9298521696825],[-72.11832498282907,43.922620750217035],[-72.12102774930786,43.91910439808352],[-72.12428242823628,43.916906458746695],[-72.13309770781052,43.913074309482205],[-72.13561403818406,43.9096204608634],[-72.14597665086772,43.90487932550794],[-72.14959474001034,43.902913994322134],[-72.15543236634507,43.897712154303335],[-72.158738093325,43.89264378348346],[-72.15992668973269,43.887741750368036],[-72.16200270417215,43.88673581315839],[-72.1696020860331,43.88620847230671],[-72.17242965650745,43.884883021208196],[-72.173700609226,43.87995479267387],[-72.17257610684821,43.876911144353464],[-72.16879982633782,43.87112802109647],[-72.16823506213646,43.86909624669278],[-72.17464666338714,43.866623179056965],[-72.18186285390934,43.86594992087055],[-72.18505444616788,43.86381845749323],[-72.18590272331913,43.85976654100671],[-72.18814343563335,43.856102385568654],[-72.18792913821976,43.85380974374856],[-72.18391627420905,43.846425356300756],[-72.18291132635584,43.84357941460322],[-72.18331339375773,43.8412246542531],[-72.18254450757644,43.83498923749615],[-72.18360278087148,43.830818825058984],[-72.18890120577946,43.82230422686891],[-72.18884410162728,43.82063708730841],[-72.18692252660539,43.81630826322559],[-72.18348206471568,43.811616717290335],[-72.18330270854617,43.80978299262562],[-72.18518850270603,43.80448364993897],[-72.19057434957688,43.80099471901339],[-72.19157086713449,43.79983346258347],[-72.19399201407825,43.7941216523629],[-72.19679703082606,43.790390299652096],[-72.20209055856144,43.787737083358074],[-72.20552273480325,43.78497461426062],[-72.20585064766199,43.78179627776054],[-72.20408091353566,43.77972535791698],[-72.20545388412975,43.776781214614],[-72.20540077140059,43.771330595300675],[-72.20846523554275,43.77216821707658],[-72.28720048632498,43.79516098497731],[-72.30512109058216,43.80048346869359],[-72.33934470947001,43.81045547150489],[-72.3461038994287,43.81160567984825],[-72.40506710826833,43.82546657927705],[-72.43831424337422,43.83228950849945],[-72.46156601403018,43.83686015250722],[-72.4595166251448,43.83810151237467],[-72.5045221614655,43.85191469303779],[-72.5453764509597,43.86434935310264],[-72.57359657753446,43.87312611162169],[-72.60255996488932,43.88178874793391],[-72.65388603721946,43.89564164550551],[-72.70930742816792,43.911252309986345],[-72.71228590372552,43.90797268103477],[-72.77350320973747,43.92454741962046],[-72.76891266149019,43.931625740039976],[-72.77514531906822,43.9336668795571],[-72.77108017968231,43.939073531628715],[-72.7899604524742,43.9453235997332],[-72.78522103663506,43.95219986860384],[-72.7917776944291,43.95397707450101],[-72.78705052498368,43.960867072488284],[-72.79104452787709,43.96225273388067],[-72.780715244428,43.97598977735712],[-72.75395914259158,44.012227211655045],[-72.74224432053903,44.029641926732175],[-72.7040941246714,44.01895617408652],[-72.68287258052305,44.012817810589866],[-72.63447677042475,44.089321151069335],[-72.62911416741265,44.08784249598819],[-72.62600172937465,44.09331555993725],[-72.61156146584501,44.11742086505553],[-72.58655925858501,44.16026882882652],[-72.56398607629137,44.154791373175435],[-72.55970923069565,44.1629214805815],[-72.55455621462724,44.161342852107026],[-72.53197999506331,44.155009386251294],[-72.52192667676394,44.15206444207164],[-72.49253917710382,44.14379296565376],[-72.47874799721224,44.140080620107454],[-72.45553959074957,44.13335300487378],[-72.44337724805202,44.12881080261438],[-72.43798635768295,44.14888072985401],[-72.42401144876301,44.20280657872514],[-72.41920040325078,44.220791279944414],[-72.36888140280671,44.20389734569612],[-72.34726079172276,44.19694973182391],[-72.30520205476473,44.18309561650894],[-72.25441585486044,44.18684745457586],[-72.19992598123842,44.19077116196788],[-72.17436575583575,44.192669627742355],[-72.17069198522198,44.19198122271944],[-72.15528201692726,44.187708317648294],[-72.04363993445739,44.15709928646535]]]},"properties":{"OBJECTID":11,"CNTY":17,"CNTYNAME":"ORANGE","ShapeSTArea":1790717634.6436005,"ShapeSTLength":208655.8853889852}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-72.36888140280671,44.20389734569612],[-72.41920040325078,44.220791279944414],[-72.42401144876301,44.20280657872514],[-72.43798635768295,44.14888072985401],[-72.44337724805202,44.12881080261438],[-72.45553959074957,44.13335300487378],[-72.47874799721224,44.140080620107454],[-72.49253917710382,44.14379296565376],[-72.52192667676394,44.15206444207164],[-72.53197999506331,44.155009386251294],[-72.55455621462724,44.161342852107026],[-72.55970923069565,44.1629214805815],[-72.56398607629137,44.154791373175435],[-72.58655925858501,44.16026882882652],[-72.61156146584501,44.11742086505553],[-72.62600172937465,44.09331555993725],[-72.62911416741265,44.08784249598819],[-72.63447677042475,44.089321151069335],[-72.68287258052305,44.012817810589866],[-72.7040941246714,44.01895617408652],[-72.74224432053903,44.029641926732175],[-72.75397791580258,44.03313236688708],[-72.80403768934362,44.04756725819828],[-72.84865120492913,44.060486702729094],[-72.85403144079973,44.06189919835406],[-72.89690216528162,44.07366753803067],[-72.91678219841,44.07189194590793],[-72.9217614619574,44.09472265562716],[-72.9254993239462,44.11953787436994],[-72.9321480945318,44.15550700138633],[-72.93407807573206,44.162832681396544],[-72.95220573089938,44.16142046923946],[-72.94199936815059,44.18246161980573],[-72.91757928190204,44.23238366831349],[-72.90208280627773,44.26361270517944],[-72.89498176046104,44.27810636324729],[-72.90595977390268,44.28056333642762],[-72.88039671983401,44.32986246096231],[-72.86125327167196,44.36690949653793],[-72.85555099714806,44.36559517079108],[-72.84935519174762,44.36297172119535],[-72.84382894658889,44.36221595927199],[-72.83588657792055,44.360181909221275],[-72.83283347714442,44.358779969756306],[-72.82875852758922,44.36037562087989],[-72.82595844086234,44.35903495682558],[-72.80237685381812,44.38858937755037],[-72.83983118375824,44.40346104201913],[-72.83331806207146,44.4118588461083],[-72.81435655050836,44.43675183361455],[-72.80320710340565,44.4515249990348],[-72.78606661677529,44.44741755689836],[-72.76556049213833,44.442814590279625],[-72.73434433301003,44.4308277503959],[-72.70582918643258,44.419462616837194],[-72.68072028336027,44.409207250476335],[-72.65434084524145,44.39830648779421],[-72.62468188578923,44.44010646811361],[-72.60524465740579,44.46731608203702],[-72.60025080674642,44.47402420205444],[-72.59728607324982,44.47939396396597],[-72.5595582810994,44.45908075347308],[-72.51236044843273,44.433913952477916],[-72.49356241633227,44.4239855935797],[-72.49324310211328,44.425837594454464],[-72.48509149867076,44.436778383469836],[-72.4847430738684,44.439144316181896],[-72.47593437169152,44.45001944988791],[-72.47447865178309,44.45225470415154],[-72.46699826029956,44.46154010093318],[-72.44734333991018,44.48765525380602],[-72.43468432989302,44.505247853204025],[-72.42625290392988,44.50298755253065],[-72.38821084532559,44.4883454231688],[-72.3259063133564,44.46450883247986],[-72.3213952644,44.46317534190703],[-72.31743310412898,44.46110216260642],[-72.26316084321657,44.44031416216949],[-72.22380692971156,44.42489739713071],[-72.22314295816591,44.4238097230508],[-72.22892643882552,44.421717517721575],[-72.22720439766418,44.41939882013177],[-72.22521981200128,44.42169084186523],[-72.22193291865825,44.422844640791496],[-72.22011650170046,44.421170785733445],[-72.22204270867296,44.4142496015238],[-72.2170795103801,44.41131975338504],[-72.24406847917456,44.37504605920377],[-72.27009732807778,44.340381969881975],[-72.27688092354514,44.33424160917313],[-72.31683527392694,44.298790099871],[-72.31665271858014,44.27223731016606],[-72.33509504348709,44.24752419796214],[-72.33687028611274,44.24622356994914],[-72.35060906355639,44.2282211421443],[-72.36888140280671,44.20389734569612]]]},"properties":{"OBJECTID":12,"CNTY":23,"CNTYNAME":"WASHINGTON","ShapeSTArea":1800497340.6928558,"ShapeSTLength":212985.38951793028}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-72.86864799426148,43.300537763988615],[-72.89887414039559,43.30172192229425],[-72.94149213106009,43.30300197531988],[-72.97529615622253,43.30416746250894],[-72.97588639164515,43.29055941838742],[-72.99232803501846,43.290724407201495],[-72.98759097847974,43.30501359287805],[-72.99683647526814,43.30518272719097],[-73.04679880843494,43.30706793879609],[-73.09256990763464,43.30693898303623],[-73.13312771038603,43.30935743022102],[-73.16816682242377,43.31136774280403],[-73.21942654824204,43.31302862644884],[-73.2441024821414,43.31395126485172],[-73.25461006970804,43.31450665987397],[-73.2539160168508,43.3319012101891],[-73.25338277781049,43.35029738034275],[-73.25269732495165,43.36762893157568],[-73.25091659690938,43.40413168975246],[-73.24923165419173,43.44805395386362],[-73.24889093849576,43.46186172675026],[-73.24788116270575,43.48360364560833],[-73.24678110856827,43.51558538117965],[-73.24751086088895,43.519445531723775],[-73.24670783894528,43.521277971724864],[-73.24754959568197,43.523083551437104],[-73.24687962972318,43.52576823067947],[-73.24227496101459,43.52886492811786],[-73.2433000337073,43.53043874256001],[-73.24148788791359,43.53280956380982],[-73.24347117952522,43.53557896077211],[-73.24572678495116,43.54169560774843],[-73.25062541258357,43.544038711721804],[-73.24905236815411,43.54552157779897],[-73.25063694593939,43.55022315915025],[-73.24879456929987,43.551565484028785],[-73.24952277255105,43.55420427603413],[-73.25263045030967,43.55668601828317],[-73.25800160471016,43.56429840488268],[-73.26040616951016,43.56641465450493],[-73.26409350520063,43.56752417946615],[-73.26463965596231,43.5692954999267],[-73.26819650307773,43.56924465228636],[-73.26802382047093,43.57228441564745],[-73.2741386049287,43.572881333935406],[-73.27717168320449,43.574635635145384],[-73.28108265388995,43.574124734681476],[-73.28119587670999,43.57726157057859],[-73.28515460913228,43.579463062261475],[-73.28979390780366,43.57787845813166],[-73.29190993116951,43.57873951293353],[-73.2957937664627,43.57800957660378],[-73.29733251760445,43.5793651843621],[-73.29557849167907,43.58199836191406],[-73.29242519667578,43.58463689889784],[-73.29712483487066,43.587609516872085],[-73.29322593865692,43.592782324258316],[-73.29272380228824,43.597595759699],[-73.29101682290217,43.59923313249158],[-73.2928624505913,43.601674780246704],[-73.29250989113154,43.60352869285091],[-73.29826653719599,43.609933038687345],[-73.30238385427478,43.611973407525156],[-73.30063932536461,43.61812058183752],[-73.30318946036837,43.620351899741586],[-73.30182967245325,43.62436679500072],[-73.30265331566696,43.62576783665085],[-73.30608332497387,43.62808831343823],[-73.30748524964045,43.62787500144966],[-73.31066890901175,43.62401363617968],[-73.31376597795565,43.625896195588396],[-73.31668798419288,43.62674114168962],[-73.31814694411625,43.6285362643688],[-73.32050590689916,43.62733755485725],[-73.32374935607565,43.627754365805664],[-73.32791699205096,43.62587909815846],[-73.3309597580215,43.62594153727457],[-73.33307907012055,43.627629058281094],[-73.33681349659592,43.6266035956514],[-73.3384335689443,43.62504179096235],[-73.34138351052653,43.62458987900564],[-73.34328016280026,43.626508279448906],[-73.34668036572488,43.624666789033554],[-73.34926984248304,43.62152530004372],[-73.35186646452254,43.62276959857524],[-73.35383985696299,43.625967162349816],[-73.35742310704255,43.625748714815614],[-73.35927904344102,43.62461409240035],[-73.3661328989026,43.62337057370992],[-73.36917737905792,43.624908199455916],[-73.3718915650847,43.62457790296786],[-73.37248674957867,43.62203874345866],[-73.37007660033657,43.61927596279648],[-73.37463781757447,43.61500097487412],[-73.37648590697883,43.61256920765653],[-73.37486989656615,43.610637505627416],[-73.37463366769576,43.60817579536168],[-73.37283265828475,43.60489076745973],[-73.37468707704915,43.60262108150748],[-73.37780272400924,43.600396501866975],[-73.37726950862049,43.598570820152105],[-73.38227983922572,43.598263136196],[-73.38393037210184,43.59647722524553],[-73.38165859966058,43.59253719899186],[-73.38345706141754,43.591247897798354],[-73.38629283685832,43.5913960628456],[-73.38684068331757,43.58862136023908],[-73.38239396737232,43.586181532204044],[-73.38558481667806,43.58317497214048],[-73.38292615201833,43.57902184800732],[-73.38397468976781,43.57552600762799],[-73.38923042939975,43.574780570485274],[-73.39186342885021,43.572917495655545],[-73.39181960287621,43.5699728485422],[-73.39931202535168,43.56819589086273],[-73.40669772544533,43.57151350209746],[-73.40812620055401,43.572757872121436],[-73.41653498644078,43.577367514056974],[-73.42094478838143,43.581732492911655],[-73.42688109483993,43.583020959205605],[-73.42947960537636,43.58472261964251],[-73.43133370273537,43.58792613436744],[-73.42990838064509,43.591292343141646],[-73.42765315591528,43.593989218857146],[-73.42524353505455,43.5982777416026],[-73.4230184497032,43.601056366427066],[-73.4216898719805,43.604432395570285],[-73.42383023342614,43.61250214944085],[-73.41930711674978,43.61841864472236],[-73.4178743721742,43.62214387753624],[-73.42018007123399,43.62562474900797],[-73.427470642815,43.63409827701162],[-73.42850847873098,43.63654345589168],[-73.42608409464914,43.64352420763014],[-73.42408197037133,43.64554037440247],[-73.41920143730789,43.6476615054288],[-73.4154401818473,43.65273209795381],[-73.41436681808376,43.658713387015226],[-73.41264083056052,43.66091345431829],[-73.4080256473682,43.670078636132615],[-73.40776153289701,43.67250543169679],[-73.40442452922574,43.67989615906742],[-73.40345857291422,43.684690360129004],[-73.40537476855891,43.68858050010923],[-73.40247521672568,43.69282555411705],[-73.39780470427156,43.694812689759935],[-73.39420492217897,43.698408770508266],[-73.39173995606339,43.703524108230674],[-73.38923965118488,43.70599176426905],[-73.38809112298904,43.70861358101907],[-73.3855467921182,43.71174754364005],[-73.37927062922866,43.716421424045876],[-73.37045214895797,43.72506543971701],[-73.36958438770903,43.72897030169378],[-73.37046037344662,43.735248577560036],[-73.36954384558196,43.74375292689856],[-73.36677610843313,43.747008581845215],[-73.36465444436976,43.751038912311266],[-73.36282695474505,43.75309315438833],[-73.30674037982767,43.757840554648574],[-73.25219146518187,43.76261107332067],[-73.24509457702015,43.76315070206975],[-73.20632292607849,43.766781825981404],[-73.21281149622538,43.80308167532476],[-73.21616352730264,43.821067025728574],[-73.21895880551341,43.83730054558403],[-73.1556707332272,43.83967496405282],[-73.15638742500558,43.84239222828964],[-73.12573660463974,43.8436271214489],[-73.10089415459996,43.84438633782652],[-73.09778951683998,43.84480638066194],[-73.08427769319601,43.84526985139753],[-73.07616561737193,43.84527602633414],[-73.06145308262829,43.84590098367914],[-73.03872638887616,43.8471987997],[-73.02990288913665,43.84717228045825],[-73.0300895474345,43.84493804096704],[-73.03308387609339,43.844198045351085],[-73.03080574056575,43.8413905142562],[-73.02684557542719,43.842148137862395],[-73.02227300610554,43.82438894968974],[-73.01709071336823,43.80486836039392],[-72.98943571171263,43.81461979396538],[-72.96751528233902,43.823240716074935],[-72.95974533334999,43.826085702498936],[-72.92847207760501,43.837960335777495],[-72.92723289471468,43.838732715703124],[-72.91037961632605,43.8445630599101],[-72.89147479498777,43.851572609658845],[-72.87449918314891,43.8580692509895],[-72.87387230912563,43.839637062195244],[-72.86138297446533,43.83623700187526],[-72.83834275094526,43.82926312363365],[-72.83066668842413,43.839128524018825],[-72.82238830706459,43.83646013277515],[-72.79692532178815,43.8291670848141],[-72.79541088059247,43.82563480331924],[-72.79129210190986,43.82260320501957],[-72.79372599007979,43.82206234543868],[-72.79423967863556,43.8203221256834],[-72.79211903753617,43.81651044799379],[-72.78853304424172,43.81606233136843],[-72.7844032084616,43.81384073436037],[-72.7843516385832,43.81109282292592],[-72.787848756436,43.8089681345466],[-72.78869832705348,43.80730809401689],[-72.7875111310777,43.805715292370635],[-72.78285746796037,43.80401599297043],[-72.79782246707592,43.77938827783577],[-72.81926310468323,43.74632159535934],[-72.80392886112465,43.74098192214516],[-72.81998614127639,43.71670554625107],[-72.6997966587968,43.68458556729994],[-72.72756494419221,43.63618145870498],[-72.7305049194344,43.63074079085405],[-72.75153983430204,43.59801341265904],[-72.75794942164497,43.5874325129525],[-72.77307441914257,43.591446795356354],[-72.78222425102138,43.54323446733776],[-72.78585866267535,43.5249825158868],[-72.79172429123342,43.4970631502777],[-72.77254451909191,43.49220956164092],[-72.77474320844217,43.47936656459434],[-72.75402169880877,43.474900982084804],[-72.72422431737083,43.46818253753719],[-72.72736822235086,43.46107089244957],[-72.72256418430345,43.46008526258974],[-72.73840418568807,43.4170448480358],[-72.75522264269321,43.384725863370484],[-72.762356028391,43.37143029541823],[-72.7584914379173,43.35331542836739],[-72.78361960585687,43.36392427189571],[-72.80059721691116,43.37143032610788],[-72.82193521746565,43.380605422746356],[-72.82198340435433,43.375719072315235],[-72.85238506315532,43.37628742808133],[-72.85470147202362,43.349286212713345],[-72.86497354118751,43.348700497881545],[-72.86864799426148,43.300537763988615]]]},"properties":{"OBJECTID":13,"CNTY":21,"CNTYNAME":"RUTLAND","ShapeSTArea":2447568361.554348,"ShapeSTLength":268308.45452096255}},
+  {"type":"Feature","geometry":{"type":"Polygon","coordinates":[[[-72.95220573089938,44.16142046923946],[-72.93407807573206,44.162832681396544],[-72.9321480945318,44.15550700138633],[-72.9254993239462,44.11953787436994],[-72.9217614619574,44.09472265562716],[-72.91678219841,44.07189194590793],[-72.89690216528162,44.07366753803067],[-72.85403144079973,44.06189919835406],[-72.84865120492913,44.060486702729094],[-72.80403768934362,44.04756725819828],[-72.75397791580258,44.03313236688708],[-72.74224432053903,44.029641926732175],[-72.75395914259158,44.012227211655045],[-72.780715244428,43.97598977735712],[-72.79104452787709,43.96225273388067],[-72.80350464771345,43.945672456924875],[-72.80999702540517,43.93609650849612],[-72.81632926346374,43.937815227371466],[-72.83831069591231,43.90376684828651],[-72.86127932847306,43.86774793317464],[-72.86529107810452,43.8616561942865],[-72.87417710995096,43.86407016432933],[-72.87534563607142,43.86770511840981],[-72.87448302748876,43.872379171063066],[-72.87655389178019,43.902772931878175],[-72.9027925912886,43.89283217705257],[-72.92762592253814,43.88432468484837],[-72.9454187973384,43.8775584334656],[-72.9531899187358,43.87486735658826],[-72.96974979784696,43.87967762868503],[-72.97404303754013,43.881447290944294],[-72.98015118570738,43.88293286757756],[-72.98023112484987,43.86397512998485],[-72.97289506721809,43.86677686576626],[-72.96551127410675,43.844602358821994],[-72.95974533334999,43.826085702498936],[-72.96751528233902,43.823240716074935],[-72.98943571171263,43.81461979396538],[-73.01709071336823,43.80486836039392],[-73.02227300610554,43.82438894968974],[-73.02684557542719,43.842148137862395],[-73.03080574056575,43.8413905142562],[-73.03308387609339,43.844198045351085],[-73.0300895474345,43.84493804096704],[-73.02990288913665,43.84717228045825],[-73.03872638887616,43.8471987997],[-73.06145308262829,43.84590098367914],[-73.07616561737193,43.84527602633414],[-73.08427769319601,43.84526985139753],[-73.09778951683998,43.84480638066194],[-73.10089415459996,43.84438633782652],[-73.12573660463974,43.8436271214489],[-73.15638742500558,43.84239222828964],[-73.1556707332272,43.83967496405282],[-73.21895880551341,43.83730054558403],[-73.21616352730264,43.821067025728574],[-73.21281149622538,43.80308167532476],[-73.20632292607849,43.766781825981404],[-73.24509457702015,43.76315070206975],[-73.25219146518187,43.76261107332067],[-73.30674037982767,43.757840554648574],[-73.36282695474505,43.75309315438833],[-73.354648445764,43.764572725502596],[-73.35126754951067,43.76983538100686],[-73.35100010321158,43.77244303171909],[-73.35536631963413,43.77793478943567],[-73.35752352897394,43.785565475457155],[-73.36221452812659,43.78996275775821],[-73.36810719556995,43.79334275459938],[-73.3767949231933,43.79916528330683],[-73.37789099651526,43.804889196223854],[-73.37959879194007,43.8092079105004],[-73.3820612560783,43.812223084095315],[-73.38956261329233,43.81704661820991],[-73.3921767980732,43.82047791980243],[-73.39279775360109,43.82278250002241],[-73.39158515804733,43.82653786934774],[-73.38930148959587,43.8312422009364],[-73.38500565392424,43.83535014388469],[-73.37779170999922,43.83872343801077],[-73.37557132416589,43.840555860872314],[-73.37258901782285,43.84536600399219],[-73.37363106197103,43.84753236627708],[-73.38025361093257,43.851930454717056],[-73.38200628687048,43.85412465322361],[-73.38219049129683,43.857026765314544],[-73.37909093967995,43.86519422026742],[-73.374345815465,43.874319133974964],[-73.37415185331291,43.876243190317865],[-73.37804105163684,43.882915864355],[-73.38138127823294,43.88752785985118],[-73.38754990018056,43.896996912046085],[-73.39005925213148,43.89940919777934],[-73.39589336453656,43.902984353336564],[-73.39977442446914,43.91215039899103],[-73.40206475929526,43.919249614413665],[-73.40712600172418,43.92819806327717],[-73.40886367625122,43.934515782821805],[-73.40745181837167,43.94249126648459],[-73.40565047127826,43.948074928454254],[-73.40572850005542,43.95593595079858],[-73.40644715945882,43.95959245562314],[-73.40707289178405,43.96819874497054],[-73.41204809473662,43.977613792285176],[-73.41252584187313,43.98222070067822],[-73.41120053002926,43.98628137422473],[-73.41040614540445,43.991861534815556],[-73.40906365624136,43.996020326093124],[-73.40880652680718,44.00068897148931],[-73.40754800813865,44.00437204123082],[-73.40600910228756,44.01142761428042],[-73.4061184399091,44.01638044008752],[-73.40764954686456,44.021119291786405],[-73.41039191690756,44.02651457458072],[-73.41406440369633,44.0293112975423],[-73.42145085789731,44.03186272364978],[-73.42509391981466,44.03438914220166],[-73.42689672179547,44.036700254815294],[-73.43570605295727,44.04188073900737],[-73.43790305974272,44.04513408772652],[-73.43721088246481,44.04844731040146],[-73.43465701854548,44.055308314112445],[-73.43253970635345,44.06233840220122],[-73.43023176410397,44.07194860792794],[-73.43004019528664,44.07720423033417],[-73.42921489959586,44.07983986560239],[-73.42325279996187,44.08895855282532],[-73.41857918886433,44.096646266374634],[-73.41521957110335,44.101353249755476],[-73.41393700271036,44.10785334799586],[-73.41204085684528,44.1106539182666],[-73.41122816997458,44.11377620930959],[-73.41201790511667,44.11963872364655],[-73.41405297036043,44.12408816953686],[-73.41416274702557,44.128167978891355],[-73.41613987499379,44.13236395489455],[-73.41197695632368,44.13778915290618],[-73.40812900436936,44.1399051915124],[-73.4030477704112,44.14499265553143],[-73.40214859349277,44.14806874797177],[-73.40287761236362,44.150490997689],[-73.40093607605449,44.152706005827994],[-73.39885792948176,44.16192293353991],[-73.3958105759441,44.166274131209846],[-73.39749526295682,44.170924491855864],[-73.39717544342035,44.17384666980075],[-73.39465757832305,44.17738529586088],[-73.3919109396608,44.17843636460983],[-73.390232908645,44.18033943613627],[-73.38998295339124,44.18378384961596],[-73.39090971424763,44.188118089451685],[-73.39062792668722,44.191075853614386],[-73.38889155518831,44.19262737048839],[-73.38592980096331,44.19277475745279],[-73.38376004867972,44.19404065968548],[-73.38193651511551,44.19774209787387],[-73.37373514356533,44.200974882576666],[-73.37086373638103,44.20453448223327],[-73.36339202175441,44.20774142640141],[-73.36184988929215,44.20942983800203],[-73.36222543261894,44.212468959355476],[-73.35817784361377,44.21636299050364],[-73.35566009632157,44.22006371945068],[-73.35537549590265,44.223165309966554],[-73.35138414446683,44.225988771056905],[-73.35047837617378,44.22990046008063],[-73.34951277013073,44.230913827281796],[-73.34393919123968,44.23370468159618],[-73.34263645772084,44.23534245872185],[-73.34337937352055,44.23819802173167],[-73.33912370423096,44.23890147031834],[-73.33521134816135,44.24048270343603],[-73.33038018505835,44.24432035019782],[-73.32465380050384,44.24352503139915],[-73.3231777538876,44.24507014710151],[-73.32247997454586,44.248741698087905],[-73.31904780259022,44.25029371732127],[-73.31729544981461,44.25338321302669],[-73.31674743039527,44.25650817441441],[-73.31712142164382,44.263005241902995],[-73.31328149172086,44.26414772030446],[-73.27038317174949,44.26011511493143],[-73.24914345799375,44.26177231255189],[-73.19679836287507,44.266642030713605],[-73.16183248633514,44.271536612655844],[-73.15219676486767,44.27245979434792],[-73.10473136432508,44.278842321269124],[-73.08439116483443,44.28171859313689],[-73.07058802229065,44.2838702930658],[-73.07001238401223,44.279934866078236],[-73.01651933434506,44.28536585557466],[-73.02037284201288,44.30375376633077],[-73.0058479191561,44.3009831930459],[-72.97588174197186,44.29504739217094],[-72.97080187807347,44.26334646181387],[-72.96545080319989,44.23139972811686],[-72.95703444625443,44.185816672730425],[-72.95220573089938,44.16142046923946]]]},"properties":{"OBJECTID":14,"CNTY":1,"CNTYNAME":"ADDISON","ShapeSTArea":2091812272.7567062,"ShapeSTLength":227767.352824913}}
+  ]}
+  
+  export default countyData;
