+++ src/bordervt.js
@@ -1,0 +1,98 @@
+const borderData = {
+  "type": "Feature",
+  "properties": {
+    "name": "Vermont",
+    "name_alt": "VT",
+    "type": "State",
+    "latitude": 44.0886,
+    "longitude": -72.7317
+  },
+  "geometry": {
+    "type": "Polygon",
+    "coordinates": [
+      [
+        [-73.35218221090553, 45.00541896831666],
+        [-73.18201496808604, 45.00562770855899],
+        [-72.76588580182134, 45.00611110701493],
+        [-72.34976762188519, 45.006605491799434],
+        [-71.9336384556205, 45.00706691759828],
+        [-71.51752027568435, 45.007561302382754],
+        [-71.53348341105892, 44.98797267858865],
+        [-71.51022535353107, 44.90834376930104],
+        [-71.62065992805631, 44.77189356878338],
+        [-71.6182868810908, 44.72776148702175],
+        [-71.56844190848624, 44.60763697072062],
+        [-71.57149610782149, 44.57919336611994],
+        [-71.58779981938093, 44.565262701526024],
+        [-71.60928907801318, 44.514055424182004],
+        [-71.63820509474124, 44.481678713962566],
+        [-71.68300734254406, 44.45027978698354],
+        [-71.82551101008951, 44.37408959853025],
+        [-71.86767653904175, 44.3549184552208],
+        [-72.00177566525097, 44.32949609096974],
+        [-72.0312080394206, 44.300733882841314],
+        [-72.06223343122913, 44.11635033088575],
+        [-72.10621170439114, 44.013078842571275],
+        [-72.11492386292662, 43.965431135675516],
+        [-72.17385452923715, 43.884088359135035],
+        [-72.22238114241641, 43.79086936143841],
+        [-72.29669266868866, 43.7149538311987],
+        [-72.33767167415814, 43.621987519058564],
+        [-72.36236894072528, 43.58664450013222],
+        [-72.38495683221193, 43.52922994716076],
+        [-72.4070723115712, 43.3320363361228],
+        [-72.43818559400808, 43.224644974604274],
+        [-72.47371538051968, 43.038536569067304],
+        [-72.49647905326304, 42.99166889149819],
+        [-72.51941850726311, 42.966674994060384],
+        [-72.54980669201608, 42.886683535930814],
+        [-72.55280595970862, 42.85644915977747],
+        [-72.5396223654557, 42.82963153180134],
+        [-72.48365800785209, 42.76617449813065],
+        [-72.46685991150817, 42.730315121762715],
+        [-72.56517656564927, 42.733050717570194],
+        [-72.66349321979038, 42.73578631337767],
+        [-72.76177691494586, 42.738532895513686],
+        [-72.86009356908696, 42.741268491321165],
+        [-72.95841022322807, 42.744004087128644],
+        [-73.05669391838354, 42.74673968293612],
+        [-73.15501057252465, 42.749486265072164],
+        [-73.25332722666576, 42.752221860879644],
+        [-73.28019978628461, 42.81315203898518],
+        [-73.26655476623284, 42.864776796813885],
+        [-73.26454426810928, 42.93521014911008],
+        [-73.26192952191579, 43.027066842067256],
+        [-73.25908406282286, 43.1278554201308],
+        [-73.25523884783243, 43.26537129451725],
+        [-73.25265706062457, 43.357195028488775],
+        [-73.25032795897322, 43.439669396869306],
+        [-73.24709797838126, 43.55304830744436],
+        [-73.3061604806343, 43.616252655558554],
+        [-73.33048421203094, 43.626689667675436],
+        [-73.3492378748557, 43.620383515091135],
+        [-73.39143636279358, 43.578525603338136],
+        [-73.40494954690281, 43.57881124788028],
+        [-73.41592488911837, 43.589643767824754],
+        [-73.4133650745676, 43.62785421850111],
+        [-73.3758467625895, 43.76218405760977],
+        [-73.37496785630599, 43.804151832648216],
+        [-73.38300984880026, 43.875826640069874],
+        [-73.42494466485307, 44.038292466579975],
+        [-73.42613118833583, 44.074854967974716],
+        [-73.40975057247658, 44.131412587319716],
+        [-73.38112020029067, 44.18743187656605],
+        [-73.33182454411332, 44.24458275765244],
+        [-73.32091511986903, 44.268873530063445],
+        [-73.32463948524548, 44.367860350245735],
+        [-73.30975301006823, 44.4596401389031],
+        [-73.37207845189889, 44.59733179454625],
+        [-73.34856770881451, 44.77529933063204],
+        [-73.3658052583002, 44.860366472548975],
+        [-73.34505208368041, 44.93886378999653],
+        [-73.35218221090553, 45.00541896831666]
+      ]
+    ]
+  }
+}
+
+export default borderData;
